% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{as_historty_array}
\alias{as_historty_array}
\title{Changes an array of bases into a "history_array" for use in \code{tourr::interpolate}}
\usage{
as_historty_array(basis_array, data)
}
\arguments{
\item{basis_array}{An array of bases.}

\item{data}{The data matrix to be projected through the basis. This is
\code{tourr::save_history} objects, but not consumed downstream in \code{spinifex}.}
}
\value{
An array of numeric bases with custom class "history_array" for
consumption by \code{tourr::interpolate}.
}
\description{
Attaches data
}
\examples{
rb <- tourr::basis_random(6, 2)
pan_zoom(pan = c(-1, 0), zoom = c(2/3, 2/3), x = rb)
}
\seealso{
\code{\link[tourr]{save_history}} for preset choices.
}
