% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{basis_half_circle}
\alias{basis_half_circle}
\title{Create a basis that gives uniform contribution in a circle}
\usage{
basis_half_circle(data)
}
\arguments{
\item{data}{The data to create a basis for.}
}
\description{
Orthonormalizes uniform variable contributions on a unit circle. This
serves as a NULL basis, one that is variable agnostic while spacing the
variables to have minimize variable dependence.
}
\examples{
dat <- scale_sd(wine[, 2:6])
bas <- basis_half_circle(dat)
}
\seealso{
Other basis producing functions: 
\code{\link{basis_guided}()},
\code{\link{basis_odp}()},
\code{\link{basis_olda}()},
\code{\link{basis_onpp}()},
\code{\link{basis_pca}()}
}
\concept{basis producing functions}
