% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{basis_identity}
\alias{basis_identity}
\title{Create an indentity basis}
\usage{
basis_identity(p, d = 2)
}
\arguments{
\item{p}{number of dimensions of the data}

\item{d}{number of dimensions of the basis, defaulting to 2}
}
\value{
[p, d=2] indentity matrix followed by 0s
}
\description{
Creates a [p, d=2] indentity basis; indentity matrix followed by 0s
}
