% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{basis_lda}
\alias{basis_lda}
\title{The basis of Linear Discriminant Analysis (LDA)}
\usage{
basis_lda(data, class, p = 2L)
}
\arguments{
\item{data}{Numeric matrix or data.frame of the observations, coereced to matrix.}

\item{class}{The class for each observation, coereced to a factor.}

\item{p}{Number of dimensions in the projected space.}
}
\value{
Numeric matrix of the last basis of a guided tour.
}
\description{
Returns a numeric matrix of the first \code{p} columns of the MASS::lda for the
given class. MASS::lda()$scaling is not orthonromal (!?); coerrced
with tourr::orthonormalise().
}
\examples{
basis_lda(data = wine[, 2:14], class = wine$Type)
}
\seealso{
\code{\link{MASS::lda}}
}
