% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{basis_olpp}
\alias{basis_olpp}
\title{The basis of Orthogonal Locality Preserving Projection (OLPP)}
\usage{
basis_olpp(data, d = 2L, ...)
}
\arguments{
\item{data}{Numeric matrix or data.frame of the observations, coerced to matrix.}

\item{d}{Number of dimensions in the projection space.}

\item{...}{Optional, other arguments to pass to \code{\link[Rdimtools]{do.odp}}.}
}
\value{
Orthogonal matrix basis that distinguishes the levels of \code{class}
based on local and non-local variation as weighted against the neighborhood
graph.
}
\description{
Orthogonal Locality Preserving Projection (OLPP) is the orthogonal variant of
LPP, a linear approximation to Laplacian Eigenmaps. It finds a linear
approximation to the eigenfunctions of the Laplace-Beltrami operator on the
graph-approximated data manifold.
}
\examples{
dat_std <- scale_sd(wine[, 2:14])
basis_olpp(data = dat_std)
}
\references{
He X (2005). Locality Preserving Projections. PhD Thesis,
University of Chicago, Chicago, IL, USA.
}
\seealso{
\code{\link[Rdimtools]{do.odp}}

Other {basis identifiers}: 
\code{\link{basis_guided}()},
\code{\link{basis_odp}()},
\code{\link{basis_olda}()},
\code{\link{basis_onpp}()},
\code{\link{basis_pca}()}
}
\concept{{basis identifiers}}
