% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{basis_pca}
\alias{basis_pca}
\title{The basis of Principal Component Analysis (PCA)}
\usage{
basis_pca(data, p = 2L)
}
\arguments{
\item{data}{Numeric matrix or data.frame of the observations.}

\item{p}{Number of dimensions in the projected space.}
}
\description{
The basis of Principal Component Analysis (PCA)
}
\examples{
basis_pca(data = wine[, 2:14])
}
