% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{basis_random}
\alias{basis_random}
\title{Create a random basis}
\usage{
basis_random(p, d = 2)
}
\arguments{
\item{p}{number of dimensions of the data}

\item{d}{number of dimensions of the basis.Defaults to 2.}
}
\value{
orthonormalized matrix, the [p, d=2] basis
}
\description{
Creates a [p, d=2] random orthonormal basis
}
