% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{col_of}
\alias{col_of}
\title{Return hex color code for a given discrete categorical variable.}
\usage{
col_of(class, pallet_name = "Dark2")
}
\arguments{
\item{class}{The discrete categorical variable to return the color of.}

\item{pallet_name}{The name of the \code{RColorBrewer} pallet to get the colors
from. Defaults to "Dark2".}
}
\value{
Vector of character hex color code of the passed categorical variable.
}
\description{
Return hex color code for a given discrete categorical variable.
}
\examples{
col_of(tourr::flea$species)
}
