% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manip_space.r
\name{create_manip_space}
\alias{create_manip_space}
\title{Create a manipulation space}
\usage{
create_manip_space(basis, manip_var)
}
\arguments{
\item{basis}{[p, d=2] orthonormal basis}

\item{manip_var}{number of the variable to rotate}
}
\value{
manip_space, a [p, d+1=3] orthonormal manipulation space
}
\description{
Primarily internal use. Create an [p, d+1=3] orthonormal manipulation space from the given basis concatonated with a zero vector, with manip_var set to 1.
}
