% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_basis.r
\name{create_random_basis}
\alias{create_random_basis}
\title{Create and return an orthonormalized random basis}
\usage{
create_random_basis(p, d = 2)
}
\arguments{
\item{p}{number of dimensions of the data}

\item{d}{number of dimensions of the basis.Defaults to 2.}
}
\value{
orthonormalized matrix, the [p, d=2] basis
}
\description{
Creates a [p, d=2] basis of random values then orthonormalize it
}
