% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{create_slideshow}
\alias{create_slideshow}
\title{Create a slideshow array of the projected bases}
\usage{
create_slideshow(tour, data = NULL)
}
\arguments{
\item{data}{[n, p] dim data to project, consisting of 
only numeric variables (for coercion into matrix.)}

\item{m_tour}{the output of manual_tour(), list of projection bases by index.}
}
\description{
Takes the result of manual_tour() and interpolated the data over the 
the tour path of the reference frame.
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])

rb <- tourr::basis_random(n = ncol(flea_std))
mtour <- manual_tour(rb, manip_var = 4)
sshow <- create_slideshow(tour = mtour, data = flea_std)
}
