% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{data_proj}
\alias{data_proj}
\title{Project data onto a rotated space}
\usage{
data_proj(data = quakes, r_space)
}
\arguments{
\item{data}{[n, p] data to project, consisting of only numeric variables (for coercion into matrix)}

\item{r_space}{[p, 3] rotated manipulation space}
}
\description{
Project [n, p] data onto a [p, 3] rotated manipulation space into the [n, 3] projection. May be used after manipulations
}
\examples{
this_r_space <- horizontal_manip(manip_space = this_manip_space, phi = i)
proj <- data_proj(data = quakes, r_space = this_r_space)
plot(proj[,1], proj[,2], main="Projected data")

for (i in seq(0, pi, pi/20)) {
this_r_space <- horizontal_manip(manip_space = this_manip_space, phi = i)
proj <- data_proj(data = quakes, r_space = this_r_space)
plot(proj[,1], proj[,2], main="Projected data")
Sys.sleep(time=.5)
print(i/pi)
}

}
