% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{devMessage}
\alias{devMessage}
\title{Development message}
\usage{
devMessage(text)
}
\arguments{
\item{text}{A character string to message() if package version is _9000.}
}
\description{
Send a message if the 4th chunk of the package version is 9000.
}
