% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{.bind2data_df}
\alias{.bind2data_df}
\title{Binds replicated elements of aes_args list to df_data}
\usage{
.bind2data_df(list)
}
\arguments{
\item{list}{A list of arguments such as those passed in \code{aes_args} and
\code{identity_args}.}

\item{nrow_frames}{Scalar number of rows in the data frames to replicate to.}

\item{nrow_data}{Scalar number of rows in the data to replicate.}
}
\description{
Internal function. To be applied to \code{aes_args}
replicates elements to length data
}
\examples{
## This function is not meant for external use
}
\seealso{
Other Internal utility: 
\code{\link{.init4proto}},
\code{\link{lapply_rep_len}()}
}
\concept{Internal utility}
