% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\docType{data}
\name{.fortify_data}
\alias{.fortify_data}
\title{Prep data, especially for local data passes in \verb{proto_*}}
\format{
An object of class \code{expression} of length 1.
}
\usage{
.fortify_data
}
\description{
Internal expression. Creates local .objects to be commonly consumed by
spinifex proto_* functions.
}
\examples{
## This expression. is not meant for external use.
}
\seealso{
Other Internal utility: 
\code{\link{.bind_elements2df}()},
\code{\link{.init4proto}},
\code{\link{.lapply_rep_len}()}
}
\concept{Internal utility}
\keyword{datasets}
