% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{.init4proto}
\alias{.init4proto}
\title{Initialize common obj from .global \code{ggtour()} objects & test their existence}
\usage{
.init4proto()
}
\description{
Internal function. Creates local .objects to be commonly consumed by spinifex
proto_* functions
}
\examples{
## This function is not meant for external use.
}
\seealso{
Other Internal utility: 
\code{\link{lapply_rep_len}()}
}
\concept{Internal utility}
