% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{.lapply_rep_len}
\alias{.lapply_rep_len}
\title{Replicate all vector elements of a list}
\usage{
.lapply_rep_len(list, to_length, expected_length)
}
\arguments{
\item{list}{A list of arguments such as those passed in \code{aes_args} and
\code{identity_args}.}

\item{to_length}{Scalar number, length of the output vector;
the number of rows in the data frames to replicate to.}

\item{expected_length}{Scalar number, the expected length of the each element
of \code{list}.}
}
\description{
Internal function. To be applied to \code{aes_args} and \code{identity_args},
replicates vectors of length data to length of data*frames for animation.
}
\examples{
## This function is not meant for external use
}
\seealso{
Other Internal utility: 
\code{\link{.bind_elements2df}()},
\code{\link{.fortify_data}},
\code{\link{.init4proto}}
}
\concept{Internal utility}
