% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_basisAxes.R
\name{geom_basisAxes}
\alias{geom_basisAxes}
\title{Depicts the axes contribution described by the basis}
\usage{
geom_basisAxes(
  basis = NULL,
  manip_var = NULL,
  manip_col = "blue",
  theta = 0,
  phi = 0,
  lab = NULL,
  x_offset = -1,
  y_offset = 0,
  scale = 2/3,
  ...
)
}
\arguments{
\item{basis}{A (p, d) dim orthonormal numeric matrix.
Defaults to NULL, giving a random basis.}

\item{manip_var}{Number of the variable to rotate,
defaults to NULL, no variable highlighted.}

\item{theta}{Angle in radians of "in-projection plane" rotation,
on the XY plane of the reference frame. Defaults to 0, no rotaion.}

\item{phi}{Angle in radians of the "out-of-projection plane" rotation, into
the z-direction of the axes. Defaults to 0, no rotaion.}

\item{x_offset}{Numeric value to pan in the x-direction. Defaults to -1.}

\item{y_offset}{Numeric value to pan in the x-direction. Defaults to 0.}

\item{scale}{Numeric value to scale/zoom the size for x. Defaults to 2/3.}

\item{...}{Optionally pass additional arguments. ie: pch, col, cex}
}
\value{
a ggplot object of the rotated projection.
}
\description{
Composite ggplot2 geom_ object (ggproto) that illustrated variable
directions and magnitudes of the projection basis.
}
\examples{
rb <- tourr::basis_random(n = ncol(flea_std))

ggplot2::ggplot() + 
  geom_basisAxes(basis = rb, manip_var = 2, manip_col = "purple")
}
