% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_manualTour_frame.r
\name{geom_manualTour_frame}
\alias{geom_manualTour_frame}
\title{geom_ object of a single projection frame of a manual tour}
\usage{
geom_manualTour_frame(
  basis = NULL,
  data = NULL,
  manip_var = NULL,
  theta = 0,
  phi = 0,
  rescale_data = FALSE,
  x_offset = 0,
  y_offset = 0,
  scale = 1,
  ...
)
}
\arguments{
\item{basis}{A (p, d) dim orthonormal numeric matrix.
Defaults to NULL, giving a random basis.}

\item{data}{A (n, p) dataset to project, consisting of numeric variables.}

\item{manip_var}{Number of the variable to rotate.}

\item{theta}{Angle in radians of "in-projection plane" rotation,
on the XY plane of the reference frame. Defaults to 0, no rotaion.}

\item{phi}{Angle in radians of the "out-of-projection plane" rotation, into
the z-direction of the axes. Defaults to 0, no rotaion.}

\item{rescale_data}{When TRUE scales the data to between 0 and 1.
Defaults to FALSE.}

\item{x_offset}{Numeric value to pan in the x-direction. Defaults to 0.}

\item{y_offset}{Numeric value to pan in the y-direction. Defaults to 0.}

\item{scale}{Numeric value to scale/zoom the size for x. Defaults to 1.}

\item{...}{Optionally pass additional arguments. ie: pch, col, cex}
}
\value{
A ggplot2 geom_point object of a object of the rotated projection.
}
\description{
One static frame of manual tour.
Returns ggplot geom_point object (ggproto) projecting the specified rotation.
Useful for providing user-guided interaction.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[,1:6])
rb <- tourr::basis_random(n = ncol(flea_std))

ggplot2::ggplot() + 
  geom_manualTour_frame(data = flea_std, basis = rb, manip_var = 4)
}
