% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_ggproto_util.r
\name{ggproto_basis_axes}
\alias{ggproto_basis_axes}
\title{ggproto for the basis axes unit circle}
\usage{
ggproto_basis_axes(
  basis_frames,
  data_frames = NULL,
  manip_var = NULL,
  axes = "center",
  manip_col = "blue",
  line_size = 1L,
  text_size = 5L
)
}
\description{
Returns headless (without ggplot()) list of geoms that make up the basis
aces unit circle (namely the geoms: line, segment and text)
}
\examples{
## See use in
if(interactive())
  ?render_
}
