% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_ggproto_util.r
\name{ggproto_data_point}
\alias{ggproto_data_point}
\title{ggproto for data point projection}
\usage{
ggproto_data_point(data_frames, aes_args, identity_args)
}
\description{
Returns headless (without ggplot()) geom_point call for projected data points.
}
\examples{
## See use in
if(interactive())
  ?render_
}
