% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_basisText.r
\name{grob_basisText}
\alias{grob_basisText}
\title{Creates a grob of the specified basis}
\usage{
grob_basisText(
  basis = NULL,
  manip_var = NULL,
  manip_col = "blue",
  theta = 0,
  phi = 0,
  lab = NULL,
  x_offset = -2,
  y_offset = 0,
  scale = 2/3,
  ...
)
}
\arguments{
\item{basis}{A (p, d) dim orthonormal numeric matrix.
Defaults to NULL, giving a random basis.}

\item{manip_var}{Number of the variable to rotate,
defaults to NULL, no variable highlighted.}

\item{theta}{Angle in radians of "in-projection plane" rotation,
on the XY plane of the reference frame. Defaults to 0, no rotaion.}

\item{phi}{Angle in radians of the "out-of-projection plane" rotation, into
the z-direction of the axes. Defaults to 0, no rotaion.}

\item{x_offset}{Numeric value to pan in the x-direction. Defaults to -2.}

\item{y_offset}{Numeric value to pan in the x-direction. Defaults to 0.}

\item{scale}{Numeric value to scale/zoom the size for x. Defaults to 2/3.}

\item{...}{Optionally pass additional arguments. ie: pch, col, cex}
}
\value{
A grob of the specified basis.
}
\description{
Creates a grob of the specified basis. Can be added with
gridExtra::grid.arrange(), but not layered on like a geom_ (ggproto) object.
}
\examples{
rb <- tourr::basis_random(n = ncol(flea_std))

geom_basisText(basis = rb, manip_var = 2, manip_col = "purple")
}
