% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_types.r
\name{horizontal_manip}
\alias{horizontal_manip}
\title{Horizontaly roatate 1 dim of a p-dim basis.}
\usage{
horizontal_manip(manip_space, phi, ...)
}
\arguments{
\item{phi}{angle changing the magnitude of `manip_var`}

\item{manip_var}{number of the variable to manipulate}
}
\value{
r_space, the [3, p] x,y,z contribution of the rotated manipulation space
}
\description{
Primarily for internal use by data_proj(). Performs a horizontal rotation on 1 dimension of a p-dimensional manipulation space. Returns the x,y,z contribution from each dimension of the rotated manipulation space [3, p].
}
