% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_manual_tour.r
\name{interpolate_manual}
\alias{interpolate_manual}
\title{\subsection{OLD:}{
}}
\usage{
interpolate_manual(basis_array, manip_var, angle)
}
\arguments{
\item{basis_array}{array, of the target bases, the extrema of the walk/segments.}

\item{manip_var}{The column number of the manipulation variable.}

\item{angle}{The step size between interpolated frames, in radians.
\subsection{However, now we have the bases at the ends of the walks.}{
\subsection{I think we want to calculate back to phi and apply similarly??}{
}

}}

\item{start}{WAS PHI current}

\item{end}{WAS Phi tgt}
}
\description{
\subsection{OLD:}{
}
}
