% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{is_orthonormal}
\alias{is_orthonormal}
\title{Test if a numeric matrix is orthonormal.}
\usage{
is_orthonormal(x, tol = 0.001)
}
\arguments{
\item{x}{numeric matrix}

\item{tol}{tolerance used to test floating point differences}
}
\description{
Handles more cases than tourr::is_orthonormal(). Will yield to changlees made to tourr.
}
\examples{
is_orthonormal(tourr::basis_random(n = 6))
is_orthonormal(matrix(1:12, ncol=2))
}
