% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_orthonormal.r
\name{is_orthornormal}
\alias{is_orthornormal}
\title{Check for orthonormality of a matrix (or basis)}
\usage{
is_orthornormal(basis)
}
\arguments{
\item{basis}{a matrix to check for orthonormality}
}
\description{
For internal use mainly. returns T/F.
}
