% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{last_ggtour}
\alias{last_ggtour}
\alias{set_last_ggtour}
\title{Retrieve/set the last \code{ggtour()}}
\usage{
last_ggtour()

set_last_ggtour(value)
}
\description{
Retrieve/set the last \code{ggtour()}
}
\seealso{
\code{\link[=ggtour]{ggtour()}}
}
\keyword{internal}
