% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{last_ggtour_env}
\alias{last_ggtour_env}
\alias{.set_last_ggtour_env}
\title{Retrieve/set last \code{ggtour()} information}
\usage{
last_ggtour_env()

.set_last_ggtour_env(ggtour_list)
}
\description{
Internal information, not related to the ggplot2 object, but used in ggtour
composition by the \verb{proto_*} functions to share/set changes.
}
\seealso{
\code{\link[=ggtour]{ggtour()}}
}
\keyword{internal}
