% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{manip_var_lda}
\alias{manip_var_lda}
\title{The number of the variable that has the max/min absolute value in the first
Linear Discriminant (of LDA). Useful for setting the manip_var argument.}
\usage{
manip_var_lda(data, class, func = max)
}
\arguments{
\item{data}{Numeric matrix or data.frame of the observations, coerced to matrix}

\item{class}{The class for each observation, coereced to a factor.}

\item{func}{The function to be applied, expects max or min.}
}
\value{
Numeric matrix of the last basis of a guided tour.
}
\description{
The number of the variable that has the max/min absolute value in the first
Linear Discriminant (of LDA). Useful for setting the manip_var argument.
}
\examples{
manip_var_lda(data = wine[, 2:14], class = wine$Type)
}
\seealso{
\code{\link{MASS::lda}}
}
