% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{manip_var_pca}
\alias{manip_var_pca}
\title{The number of the variable that has the max/min absolute value in the first
Principal Component (of PCA). Useful for setting the manip_var argument.}
\usage{
manip_var_pca(data, func = max)
}
\arguments{
\item{data}{Numeric matrix or data.frame of the observations.}

\item{func}{The function to be applied, expects \code{max} or \code{min}.}

\item{p}{Number of dimensions in the projected space.}
}
\description{
The number of the variable that has the max/min absolute value in the first
Principal Component (of PCA). Useful for setting the manip_var argument.
}
\examples{
manip_var_pca(data = wine[, 2:14])
}
