% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.r
\name{play_tour}
\alias{play_tour}
\title{Render display of a provided tour path and data}
\usage{
play_tour(tour, data = NULL, render_type = render_plotly,
  cat_var = NULL, fps = 3, ...)
}
\arguments{
\item{tour}{The result of \code{tourr::save_history} or \code{manual_tour}.}

\item{data}{Optional, number of columns must match that of \code{tour}.}

\item{render_type}{Which graphics to render to. Defaults to render_plotly,
alternative use render_gganimate.}

\item{...}{Optionally pass addition arguments to \code{plotly::animation_opts}.}
}
\description{
Takes the result of \code{tourr::save_history} (or \code{manual_tour} and optionally
data, renders the tour and passed data into a plotly animation object.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[,1:6])
tpath <- tourr::save_history(flea_std, tourr::guided_tour(tourr::cmass))

play_tour(tour = tpath, data = flea_std)
}
