% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_tour_path.r
\name{play_tour_path}
\alias{play_tour_path}
\title{Render display of a provided tour path}
\usage{
play_tour_path(tour_path, data = NULL, angle = 0.05,
  render_type = render_plotly, rescale_data = FALSE, ...)
}
\arguments{
\item{tour_path}{The result of \code{tourr::save_history()} or \code{manual_tour()}.}

\item{data}{Optional, number of columns must match that of \code{tour_path}.}

\item{angle}{Target distance (in radians) between steps. Defaults to .05.}

\item{render_type}{Graphics to render to. Defaults to render_plotly,
alternative use render_gganimate.}

\item{rescale_data}{When TRUE scales the data to between 0 and 1.
Defaults to FALSE.}

\item{...}{Optionally pass additional arguments to \code{render_type}.}
}
\description{
Takes the result of \code{tourr::save_history()} or \code{manual_tour()}, interpolates
over the path and renders into a selected \code{render_type}.
}
\examples{
\dontrun{
flea_std <- rescale(tourr::flea[,1:6])
tpath    <- save_history(flea_std, tour_path = grand_tour(),max = 3)

play_tour_path(tour_path = tpath, data = flea_std, angle = .15)

play_tour_path(tour_path = tpath, data = flea_std, angle = .15, fps = 4,
  render_type = render_gganimate, col = col_of(flea$species), axes = "bottomleft")
}
}
