% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tour.r
\name{print_tour}
\alias{print_tour}
\title{Render display of a provided tour path and data}
\usage{
print_tour(tour_path, data = NULL)
}
\arguments{
\item{tour_path}{The result of tourr::save_history}

\item{data}{Optional, tour::interpolate()'s the data. p must match the 
tour_path.}
}
\description{
Takes the result of tourr::save_history and data, returns a plotly object.
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])
tpath <- tourr::save_history(flea_std, tourr::guided_tour(tourr::cmass))

print_tour(tour_path = tpath, data = flea_std)
}
