% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_data.r
\name{proj_data}
\alias{proj_data}
\title{Project data by the rotated space. Return a list of projected data and basis.}
\usage{
proj_data(data, basis = basis_identity(p = ncol(data)), manip_var,
  manip_type = NULL, theta = NULL, center = TRUE, scale = TRUE,
  phi_from = 0, phi_to = 0, n_slides = 15)
}
\arguments{
\item{data}{[n, p] data to project, consisting of only numeric variables (for coercion into matrix)}

\item{basis}{the [p, 2] orthonormal starting basis. Default to the identity basis}

\item{manip_var}{integer column or string name of the variable to manipulate. Required, does not default}

\item{manip_type}{character string of the type of manipulation to use. Defaults to "radial". Alternatively use "horizontal" or "vertical" Superseds theta}

\item{theta}{the angle that the manipvar should be rotated by in parameter space. Superseded by manip_type}

\item{phi_from}{value of phi to start the projection. Defaults to 0}

\item{phi_to}{value of phi to end the projection. Defaults to 0}

\item{n_slides}{number of slides to create for slideshow(). Defaults to 15}
}
\description{
Project [n, p] data by [p, 3] rotated manipulation space into [n, 3] projection.
}
