% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_basis1d}
\alias{proto_basis1d}
\title{Tour proto for a 1D basis axes}
\usage{
proto_basis1d(
  position = c("left", "center", "right", "bottomleft", "topright", "off"),
  manip_col = "blue",
  segment_size = 2,
  text_size = 5
)
}
\arguments{
\item{manip_col}{The color to highlight the manipulation variable with. Not
applied if the tour isn't a manual tour. Defaults to "blue".}

\item{segment_size}{Thickness of the bars used to make the axes.}

\item{text_size}{Size of the text label of the variables.}
}
\description{
Adds 1D basis axes to the animation, the direction and magnitude of
contributions of the variables to the projection space.
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
gt <- tourr::save_history(dat, grand_tour(d = 1), max_bases = 3)

ggt <- ggtour(gt, dat) +
  proto_basis1d()
\dontrun{
animate_plotly(ggt)
}

ggt2 <- ggtour(gt, dat) +
  proto_basis1d(position = "right",
                segment_size = .8, 
                text_size = 8)
\dontrun{
animate_plotly(ggt2)
}
}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis2d}()},
\code{\link{proto_data_density1d_rug}()},
\code{\link{proto_data_hex}()},
\code{\link{proto_data_origin}()},
\code{\link{proto_data_point}()},
\code{\link{proto_data_text}()}
}
\concept{ggtour proto}
