% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_basis_1d}
\alias{proto_basis_1d}
\title{#TODO: HEADER TEXT}
\usage{
proto_basis_1d(
  position = "left",
  manip_col = "blue",
  segment_size = 2,
  text_size = 5
)
}
\arguments{
\item{#TODO:}{PARAMETER TEXT}
}
\description{
#TODO: paragraph TEXT
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
clas <- tourr::flea$species
bas <- basis_pca(dat)
mv  <- manip_var_of(bas)
mt_path <- manual_tour(bas, manip_var = mv, angle = .1)

ggt <- ggtour(mt_path, dat) +
  proto_basis_1d() +
  proto_data_background() +
  proto_data_points(aes_args = list(color = clas, shape = clas),
                    identity_args = list(size = 1.5, alpha = .7))
#animate_plotly(ggt) ## plotly::ggplotly() throwing error: 
## Error in -data$group : invalid argument to unary operator
animate_gganimate(ggt)
}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis_2d}()},
\code{\link{proto_data_background}()},
\code{\link{proto_data_density1d_rug}()},
\code{\link{proto_data_hex}()},
\code{\link{proto_data_points}()},
\code{\link{proto_data_text}()}
}
\concept{ggtour proto}
