% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_basis_2d}
\alias{proto_basis_2d}
\title{#TODO: HEADER TEXT}
\usage{
proto_basis_2d(
  position = "left",
  manip_col = "blue",
  line_size = 1,
  text_size = 5
)
}
\arguments{
\item{#TODO:}{PARAMETER TEXT}
}
\description{
#TODO: paragraph TEXT
}
\examples{
#TODO: EXAMPLES

}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis_1d}()},
\code{\link{proto_data_background}()},
\code{\link{proto_data_density1d_rug}()},
\code{\link{proto_data_hex}()},
\code{\link{proto_data_points}()},
\code{\link{proto_data_text}()}
}
\concept{ggtour proto}
