% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_data_density1d_rug}
\alias{proto_data_density1d_rug}
\alias{proto_data_density,}
\alias{proto_data_density1d}
\alias{dat}
\alias{<-}
\alias{scale_sd(tourr::flea[,}
\alias{1:6])}
\alias{clas}
\alias{tourr::flea$species}
\alias{bas}
\alias{basis_pca(dat)}
\alias{gt_path}
\alias{save_history(dat,}
\alias{grand_tour(),}
\alias{max}
\alias{=}
\alias{3,}
\alias{start}
\alias{bas)}
\alias{ggt}
\alias{ggtour(gt_path,}
\alias{dat)}
\alias{+}
\alias{proto_data_density1d_rug(aes_args}
\alias{list(color}
\alias{clas,}
\alias{fill}
\alias{clas))}
\alias{proto_basis1d()}
\alias{animate_plotly(ggt)}
\title{Tour proto for data, 1D density, with rug}
\usage{
proto_data_density1d_rug(aes_args = list(), identity_args = list())
}
\arguments{
\item{aes_args}{A list of aesthetic arguments to passed to
\verb{geom_point(aes(X)}. Any mapping of the data to an aesthetic,
for example, \code{geom_point(aes(color = myCol, shape = myCol))} becomes
\code{aes_args = list(color = myCol, shape = myCol)}.}

\item{identity_args}{A list of static, identity arguments passed into
\code{geom_point()}, but outside of \code{aes()}; \code{geom_point(aes(), X)}.
Typically a single numeric for point size, alpha, or similar.}
}
\description{
Adds \code{geom_density()} and \code{geom_rug} of the projected data.
}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis1d}()},
\code{\link{proto_basis2d}()},
\code{\link{proto_data_hex}()},
\code{\link{proto_data_origin}()},
\code{\link{proto_data_point}()},
\code{\link{proto_data_text}()}
}
\concept{ggtour proto}
