% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_data_hex}
\alias{proto_data_hex}
\title{#TODO: HEADER TEXT}
\usage{
proto_data_hex(bins = 30, aes_args = list(), identity_args = list())
}
\arguments{
\item{#TODO:}{PARAMETER TEXT}
}
\description{
#TODO: paragraph TEXT
}
\examples{
#TODO: EXAMPLES

diam_sm <- diamonds[sample(nrow(diamonds), 1000),]
dat <- scale_sd(diam_sm[, c(1, 5:6, 8:10)])
clas <- diam_sm$color
bas <- basis_pca(dat)
gt_array <- save_history(dat, grand_tour(), max = 3, start = bas)

ggp <- ggtour(gt_array, dat) +
  proto_basis_2d() +
  proto_data_hex(50, list(group = clas, fill= clas) ) +
  proto_data_background()

tictoc::
animate_gganimate(ggp)
}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis_1d}()},
\code{\link{proto_basis_2d}()},
\code{\link{proto_data_background}()},
\code{\link{proto_data_density1d_rug}()},
\code{\link{proto_data_points}()},
\code{\link{proto_data_text}()}
}
\concept{ggtour proto}
