% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_data_origin}
\alias{proto_data_origin}
\title{Tour proto for data origin zero mark}
\usage{
proto_data_origin()
}
\arguments{
\item{position}{The position, to place the basis axes relative to the centered
data. Expects one of c("left", "center", "right", "bottomleft", "topright",
"off"), defaults to "left".}
}
\description{
Adds a zero mark showing the location of the origin for the central data area.
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
clas <- tourr::flea$species
bas <- basis_pca(dat)
mv <- manip_var_of(bas)
gt_path <- tourr::save_history(dat, grand_tour(), max_bases = 5)

ggt <- ggtour(mt_path, dat) +
  proto_data_origin() +
  proto_data_points()
\dontrun{
animate_plotly(ggt)
}
}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis_1d}()},
\code{\link{proto_basis_2d}()},
\code{\link{proto_data_density1d_rug}()},
\code{\link{proto_data_hex}()},
\code{\link{proto_data_points}()},
\code{\link{proto_data_text}()}
}
\concept{ggtour proto}
