% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_data_points}
\alias{proto_data_points}
\title{#TODO: HEADER TEXT}
\usage{
proto_data_points(aes_args = list(), identity_args = list())
}
\arguments{
\item{#TODO:}{PARAMETER TEXT}
}
\description{
#TODO: paragraph TEXT
}
\examples{
dat <- scale_sd(mtcars)
bas <- basis_pca(dat)
gt_array <- save_history(dat, grand_tour(), max = 3, start = bas)

gg <- ggtour(gt_array, dat) +
  proto_basis_2d() +
  proto_data_points() +
  proto_data_hex()

animate_gganimate(gg)
}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis_1d}()},
\code{\link{proto_basis_2d}()},
\code{\link{proto_data_background}()},
\code{\link{proto_data_density1d_rug}()},
\code{\link{proto_data_hex}()},
\code{\link{proto_data_text}()}
}
\concept{ggtour proto}
