% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_density_ridges}
\alias{proto_density_ridges}
\alias{proto_density1d}
\title{Tour proto for data, 1D density_ridges, with rug marks}
\usage{
proto_density_ridges(
  group_by = "",
  aes_args = list(),
  identity_args = list(scale = 0.8)
)
}
\arguments{
\item{group_by}{A factor, discrete grouping clustering for the original data.
Defaults to "", a single density ridge with no y axis text.}

\item{aes_args}{A list of aesthetic arguments to passed to
\verb{geom_point(aes(X)}. Any mapping of the data to an aesthetic,
for example, \code{geom_point(aes(color = myCol, shape = myCol))} becomes
\code{aes_args = list(color = myCol, shape = myCol)}.}

\item{identity_args}{A list of static, identity arguments passed into
\code{geom_point()}, but outside of \code{aes()}, for instance
\code{geom_point(aes(...), size = 2, alpha = .7)} becomes
\code{identity_args = list(size = 2, alpha = .7)}.}
}
\description{
Adds \code{ggridges::geom_density_ridges()} to a ggtour. Does not work with
\code{animate_plotly()}.
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
clas <- tourr::flea$species
gt_path <- save_history(dat, grand_tour(), max = 3)

## With a class grouping
ggt <- ggtour(gt_path, dat) +
  proto_density_ridges(group_by = clas)
\dontrun{
animate_gganimate(ggt) ## Doesn't work with animate_plotly().
}

## Without class grouping, basis is oversized.
ggt2 <- ggtour(gt_path, dat) +
  proto_density_ridges() +
  proto_basis1d() ## .map_to is now quite large for this
\dontrun{
animate_gganimate(ggt2) ## Doesn't work with animate_plotly().
}
}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis}()},
\code{\link{proto_default}()},
\code{\link{proto_density}()},
\code{\link{proto_hex}()},
\code{\link{proto_highlight}()},
\code{\link{proto_origin}()},
\code{\link{proto_point}()},
\code{\link{proto_text}()}
}
\concept{ggtour proto}
