% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_frame_cor}
\alias{proto_frame_cor}
\title{Tour proto for within frame correlation}
\usage{
proto_frame_cor(
  position = c(0.7, -0.1),
  aes_args = list(),
  identity_args = list(size = 4),
  ...
)
}
\arguments{
\item{position}{Vector x and y position relative to the unit data position;
(0, 1) in each direction. Defaults to c(.7, -.1).}

\item{aes_args}{A list of aesthetic arguments to passed to
\verb{geom_point(aes(X)}. Any mapping of the data to an aesthetic,
for example, \code{geom_point(aes(color = myCol, shape = myCol))} becomes
\code{aes_args = list(color = myCol, shape = myCol)}.}

\item{identity_args}{A list of static, identity arguments passed into
\code{geom_point()}, but outside of \code{aes()}, for instance
\code{geom_point(aes(...), size = 2, alpha = .7)} becomes
\code{identity_args = list(size = 2, alpha = .7)}.}

\item{...}{Optionally, pass additional arguments to
\code{\link[stats:cor]{stats::cor}}, specifying the type of
within frame correlation.}

\item{tail_size}{How long the origin mark should extended
relative to the observations. Defaults to .05, 5\% of the projection space.}
}
\description{
Adds text to the animation, the frame and its specified correlation.
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
clas <- tourr::flea$species
gt_path <- tourr::save_history(dat, grand_tour(), max_bases = 5)

ggt <- ggtour(gt_path, dat, angle = .1) +
  proto_default() +
  proto_frame_cor(c(.70, -.1))
  
\dontrun{
animate_plotly(ggt)
}
}
\seealso{
\code{\link[stats:cor]{stats::cor}}
}
