% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_frame_cor2}
\alias{proto_frame_cor2}
\title{Tour proto for frames square correlation}
\usage{
proto_frame_cor2(
  identity_args = list(size = 4),
  row_index = TRUE,
  position = c(0.7, -0.1),
  ...
)
}
\arguments{
\item{identity_args}{A list of static, identity arguments passed into
\code{geom_point()}, but outside of \code{aes()}, for instance
\code{geom_point(aes(...), size = 2, alpha = .7)} becomes
\code{identity_args = list(size = 2, alpha = .7)}.}

\item{row_index}{A numeric or logical index of rows to subset to.
Defaults to NULL, all observations.}

\item{position}{Vector x and y position relative to the unit data position;
(0, 1) in each direction. Defaults to c(.7, -.1).}

\item{...}{Optionally, pass additional arguments to
\code{\link[stats:cor]{stats::cor}}, specifying the type of
within frame correlation.}
}
\description{
Adds text to the animation, the frame and its specified correlation.
}
\examples{
library(spinifex)
dat     <- scale_sd(flea[, 1:6])
clas    <- flea$species
gt_path <- save_history(dat, grand_tour(), max_bases = 5)

ggt <- ggtour(gt_path, dat, angle = .3) +
  proto_default(list(color = clas, shape = clas)) +
  proto_frame_cor2(position = c(.5, 1.1))
\dontrun{
animate_plotly(ggt)
}
}
\seealso{
\code{\link[stats:cor]{stats::cor}}
}
