% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_point}
\alias{proto_point}
\alias{proto_points}
\title{Tour proto for data point}
\usage{
proto_point(
  aes_args = list(),
  identity_args = list(),
  row_index = TRUE,
  bkg_color = "grey80"
)
}
\arguments{
\item{aes_args}{A list of aesthetic arguments to passed to
\verb{geom_point(aes(X)}. Any mapping of the data to an aesthetic,
for example, \code{geom_point(aes(color = myCol, shape = myCol))} becomes
\code{aes_args = list(color = myCol, shape = myCol)}.}

\item{identity_args}{A list of static, identity arguments passed into
\code{geom_point()}, but outside of \code{aes()}, for instance
\code{geom_point(aes(...), size = 2, alpha = .7)} becomes
\code{identity_args = list(size = 2, alpha = .7)}.}

\item{bkg_color}{The character color by name or hexidemical to display
backgrund observations, those not identified in \code{row_index}.
Defaults to "grey80". Use FALSE or NULL to skip rendering background points.
Other aesthetic values such as shape and alpha are set adopted from
\code{aes_args} and \code{identity_args}.}

\item{data}{Optionally pass a dataframe to this proto, superceeds data
passed to \code{ggtour}. Analogous to ggplot2 where data passed into a \verb{geom_*}
is used independent of data passed to \code{ggplot}.}
}
\description{
Adds \code{geom_point()} of the projected data.
}
\examples{
library(spinifex)
dat     <- scale_sd(flea[, 1:6])
clas    <- flea$species
gt_path <- save_history(dat, grand_tour(), max_bases = 5)

ggt <- ggtour(gt_path, dat, angle = .3) +
  proto_point(aes_args = list(color = clas, shape = clas),
              identity_args = list(size = 2, alpha = .7))
\dontrun{
animate_plotly(ggt)
}

## Select/highlight observations with `row_index`
ggt <- ggtour(gt_path, dat, angle = .3) +
  proto_point(aes_args = list(color = clas, shape = clas),
              identity_args = list(size = 2, alpha = .7),
              row_index = which(clas == levels(clas)[1]),
              bkg_color = "grey80") ## FALSE or NULL to skip ploting background
\dontrun{
animate_plotly(ggt)
}
}
\seealso{
Other ggtour proto functions: 
\code{\link{facet_wrap_tour}()},
\code{\link{ggtour}()},
\code{\link{proto_basis}()},
\code{\link{proto_default}()},
\code{\link{proto_density}()},
\code{\link{proto_hex}()},
\code{\link{proto_highlight}()},
\code{\link{proto_origin}()},
\code{\link{proto_text}()}
}
\concept{ggtour proto functions}
