% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_text}
\alias{proto_text}
\title{Tour proto for data, text labels}
\usage{
proto_text(
  aes_args = list(),
  identity_args = list(nudge_x = 0.05),
  row_index = TRUE
)
}
\arguments{
\item{aes_args}{A list of aesthetic arguments to passed to
\verb{geom_point(aes(X)}. Any mapping of the data to an aesthetic,
for example, \code{geom_point(aes(color = myCol, shape = myCol))} becomes
\code{aes_args = list(color = myCol, shape = myCol)}.}

\item{identity_args}{A list of static, identity arguments passed into
\code{geom_point()}, but outside of \code{aes()}, for instance
\code{geom_point(aes(...), size = 2, alpha = .7)} becomes
\code{identity_args = list(size = 2, alpha = .7)}.}

\item{data}{Optionally pass a dataframe to this proto, superceeds data
passed to \code{ggtour}. Analogous to ggplot2 where data passed into a \verb{geom_*}
is used independent of data passed to \code{ggplot}.}
}
\description{
Adds \code{geom_text()} of the projected data.
}
\examples{
library(spinifex)
dat     <- scale_sd(flea[, 1:6])
clas    <- flea$species
bas     <- basis_pca(dat)
mv      <- manip_var_of(bas)
gt_path <- save_history(dat, grand_tour(), max_bases = 5)

ggt <- ggtour(gt_path, dat, angle = .2) +
  proto_text(list(color = clas))
\dontrun{
animate_plotly(ggt)
}

## Custom labels, subset of points
ggt2 <- ggtour(gt_path, dat) +
  proto_text(list(color = clas, size = as.integer(clas)),
             list(alpha = .7),
             row_index = 1:15)
\dontrun{
animate_plotly(ggt2)
}
}
\seealso{
Other ggtour proto functions: 
\code{\link{facet_wrap_tour}()},
\code{\link{ggtour}()},
\code{\link{proto_basis}()},
\code{\link{proto_default}()},
\code{\link{proto_density}()},
\code{\link{proto_hex}()},
\code{\link{proto_highlight}()},
\code{\link{proto_origin}()},
\code{\link{proto_point}()}
}
\concept{ggtour proto functions}
