% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_render.r
\name{render_}
\alias{render_}
\title{Prepate the ggplot object before passing to either animation package.}
\usage{
render_(
  frames,
  axes = "center",
  manip_col = "blue",
  ggtheme = ggplot2::theme_void(),
  ...
)
}
\arguments{
\item{frames}{The result of \code{array2df()}, a long df of the projected frames.}

\item{axes}{Position of the axes: "center", "bottomleft", "off", "left",
"right". Defaults to "center".}

\item{manip_col}{String of the color to highlight the \code{manip_var}, if used.
Defaults to "blue".}

\item{ggtheme}{Intended for passing  a ggplot2::theme().}

\item{...}{Optionally passes arguments to the projection points inside the
aesthetics; \code{geom_point(aes(...))}.}
}
\description{
Typically called by \code{render_plotly()} or \code{render_gganimate()}. Takes the
result of \code{array2df()}, and renders them into a ggplot2 object.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[, 1:6])
rb <- tourr::basis_random(n = ncol(flea_std))
array_frames <- manual_tour(basis = rb, manip_var = 4)
df_frames <- array2df(array = array_frames, data = flea_std)

render_(frames = df_frames)

flea_class <- tourr::flea$species
render_(frames = df_frames, axes = "bottomleft", manip_col = "purple",
        color = flea_class, shape = flea_class,
        size = 2, alpha = .5,
        ggtheme = list(ggplot2::theme_void(), ggplot2::ggtitle("My title")))
}
