% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{render_slideshow}
\alias{render_slideshow}
\title{Render a slideshow of the toured data and bases}
\usage{
render_slideshow(slides, disp_type = "plotly", manip_col = "blue",
  cat_var = NULL, slide_time = 0.3, ...)
}
\arguments{
\item{slides}{The result of `create_slides`.}

\item{disp_type}{The graphics system to use. Defaults to 'plotly'.}

\item{manip_col}{String of the color to highlight the `manip_var`.}

\item{cat_var}{Categorical variable, optionally used to set the data point 
color and shape.}

\item{slide_time}{Time to show each slide for in seconds. essentially 1/fps, 
defaults to .3 seconds.}

\item{...}{Optional, pass addition arguments into `plotly::animation_opts`.}
}
\value{
An animation in `disp_type` graphics of the interpolated data and 
the corrisponding reference frame.
}
\description{
Takes `slides`, the result of `create_slides`, and renders them as a graph 
object of the `disp_type`.
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])

rb <- tourr::basis_random(n = ncol(flea_std))
mtour <- manual_tour(basis = rb, manip_var = 4)
sshow <- create_slides(tour = mtour, data = flea_std)
(pss <- render_slideshow(slides = sshow))
}
