% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_ggproto_util.r
\name{rep_arg_ls}
\alias{rep_arg_ls}
\title{Replicate a list of args}
\usage{
rep_arg_ls(arg_ls, length.out)
}
\arguments{
\item{arg_ls}{A list of arguments that need to be replicated.}

\item{length.out}{The target length to recycle to,
typically rows of the data times the number of basis in an array.}
}
\description{
Replicates each element of a list, to a target length for animating accross
many frames.
}
\examples{
## Setup
dat_std <- scale_sd(tourr::flea[, 1:6])
clas <- tourr::flea[, 7]
t1 <- tourr::save_history(dat_std, max = 3)
ls_args <- list(col = clas, shape = clas)
tgt_len <- nrow(dat_std) * dim(t1)[3]

spinifex:::rep_arg_ls(ls_args, tgt_len)
}
