% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_manual_tour.r
\name{rotate_basis}
\alias{rotate_basis}
\title{Return the manipulation space of the specified rotation}
\usage{
rotate_basis(basis = NULL, manip_var, theta = 0L, phi = 0L)
}
\arguments{
\item{basis}{A (p, d) orthonormal numeric matrix.
The linear combination the original variables contribute to projection space.
Defaults to NULL, generating a random basis.}

\item{manip_var}{Number of the column/dimension to rotate.}

\item{theta}{Angle in radians of "in-plane" rotation, on the xy plane of the
reference frame. Required, no default.
If left NULL, will initialize the radial angle of the \code{manip_var}.}

\item{phi}{Phi is angle in radians of the "out-of-plane" rotation,
the z-axis of the reference frame. Required, no default.}
}
\value{
(p, 2) matrix of the rotated basis.
}
\description{
Rotates a basis returning (p, 3) manipulation space that projects to
\code{view_frame()}. Allows for interactive use rather than producing a whole tour.
}
\examples{
## PCA basis with no rotation
dat <- wine[, 2:14]
bas <- basis_pca(dat)
mv  <- manip_var_pca(dat)
rotate_basis(bas, mv)

## Rotate random 
p      <- ncol(bas)
rb     <- tourr::basis_random(p)
rmv    <- sample(1:p, 1)
rtheta <- runif(1, 0, 2 * pi)
rphi   <- runif(1, 0, 2 * pi)
rotate_basis(basis = rb, manip_var = rmv, rtheta, rphi)
}
