% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.r
\name{run_app}
\alias{run_app}
\title{Runs a shiny app demonstrating manual tours}
\usage{
run_app()
}
\value{
opens a local shiny app
}
\description{
Runs a local shiny app that demonstrates manual tour and comparable
traditional techniques for static projections of multivariate data sets.
}
\examples{
run_app()
}
