% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{set_axes_position}
\alias{set_axes_position}
\title{Returns the axis scale and position.}
\usage{
set_axes_position(
  x,
  axes = c("center", "bottomleft", "topright", "off", "left", "right"),
  to = data.frame(x = c(0, 1), y = c(0, 1))
)
}
\arguments{
\item{x}{Numeric data object to scale and offset.}

\item{axes}{Position of the axes: "center", "bottomleft" or "off". Defaults
to "center".}

\item{to}{output range (numeric vector of length two). Min and max that x is
scaled to.}
}
\value{
Scaled and offset \code{x} typically controling axes placement.
}
\description{
Typically called, by other functions to scale axes.
}
\examples{
rb <- tourr::basis_random(4, 2)
set_axes_position(x = rb, axes = "bottomleft")
set_axes_position(x = rb, axes = "right", to = wine[, 2:3])
}
\seealso{
\code{\link{pan_zoom}} for more manual control.
}
