% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{slideshow}
\alias{slideshow}
\title{Slide show of the projected data and basis.}
\usage{
slideshow(data, bases, ...)
}
\arguments{
\item{data}{[n, p] dim data to project, consisting of 
only numeric variables (for coercion into matrix.)}

\item{bases}{the output of manual_tour(), list of projection bases by index.}
}
\description{
Takes the result of data_proj() and uses base graphics to view each index with delay in base grahics
}
\examples{

require(tourr)
data(flea)
flea_std <- rescale(flea[,1:6])

rb <- basis_random(n = ncol(flea_std), d=2)
prj <- manual_tour(rb, manip_var=4, manip_type = "radial",
 phi_from = 0, phi_to = pi, n_slides = 20)
slideshow(flea_std, prj)

}
