% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinifex-package.r
\docType{package}
\name{spinifex}
\alias{spinifex}
\alias{spinifex-package}
\title{spinifex}
\description{
\code{spinifex} is a package that extends the package \code{tourr}.
It builds the functionality for manual tours and allows other
tours to be rendered by \code{plotly} or \code{gganimate}. Tours are a class of
dynamic linear (orthogonal) projections of numeric multivariate data from
\code{p} down to \code{d} dimensions that are viewed as an animation as \code{p}-space is
rotated. Manual tours manipulate a selected variable, exploring how they
contribute to the sensitivity of the structure in the projection. This is
particularly useful after finding an interesting basis, perhaps via a
guided tour optimizing the projection for some objective function.
}
\details{
GitHub: \url{https://github.com/nspyrison/spinifex}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nspyrison/spinifex/}
  \item Report bugs at \url{https://github.com/nspyrison/spinifex/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Spyrison \email{spyrison@gmail.com}

Authors:
\itemize{
  \item Dianne Cook [thesis advisor]
}

}
