% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b1_stepwise_tour.r
\name{stepwise_hist}
\alias{stepwise_hist}
\title{The last basis of a guided tour}
\usage{
stepwise_hist(
  basis = NULL,
  data = NULL,
  measure = stats::var,
  curr_dim = ncol(data),
  decreasing = TRUE
)
}
\arguments{
\item{basis}{A full (p x 2) basis to start removing variables from.}

\item{data}{The data to project.}

\item{measure}{A univariate function to apply to each of the variables of
the data. Defaults to the variance, stats::var.}

\item{curr_dim}{The target number of variables left with non-zero
contributions. Defaults to the full dimensionality, ncol(data).}

\item{decreasing}{Whether or not the measure should be in decreasing order.
Defaults to TRUE. Removing a variable will operate in the opposite order.}
}
\value{
A list, containing the data, measure table, basis list, and current
dimensionality. This is the object tracks the current and previous history of
the stepwise tour.
}
\description{
The last basis of a guided tour
}
\examples{
dat <- wine[, 2:14]
bas <- basis_pca(dat)

sw_hist <- stepwise_hist(basis = bas, data = dat)
str(sw_hist)

sw_hist <- stepwise_hist(basis = bas, data = dat, measure = median,
                         curr_dim = 5, decreasing = FALSE)
str(sw_hist)
}
\seealso{
\code{\link[spinifex]{stepwise_add}}
\code{\link[spinifex]{stepwise_remove}}
}
