% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b1_stepwise_history.r
\name{stepwise_history}
\alias{stepwise_history}
\title{Creates the array of target basis for a stepwise tour}
\usage{
stepwise_history(
  basis = NULL,
  data = NULL,
  measure = stats::var,
  decreasing = TRUE
)
}
\arguments{
\item{basis}{A full (p x 2) basis to start removing variables from.}

\item{data}{The data to project.}

\item{measure}{A univariate function to apply to each of the variables of
the data. Defaults to variance, stats::var.}

\item{decreasing}{Whether or not the measure should be in decreasing order.
Defaults to TRUE. Removing a variable will operate in the opposite order.}
}
\value{
A basis array containing the target bases of a stepwise tour,
data and measure table are included as attributes. Given the custom classes
"history_array" for use by \code{tourr::interpolate} and "stepwise_array" to
differentiate from returns of \code{tourr::save_history}.
}
\description{
The basis array starts with the supplied \code{basis} and removes the variable
with the next lowest contribution to the univariate \code{measure}.
}
\examples{
dat_std <- scale_sd(wine[, 2:14])
bas <- basis_pca(dat)

sw_path <- stepwise_history(basis = bas, data = dat_std)
str(sw_path)

sw_path <- stepwise_history(basis = bas, data = dat_std,
                            measure = e1071::skewness, decreasing = FALSE)
str(sw_path)

\dontrun{
  play_tour_path(tour_path = sw_path, data = dat_std)
}
}
\seealso{
\code{\link[spinifex]{stepwise_add}}
\code{\link[spinifex]{stepwise_remove}}
}
