% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{vertical_manip}
\alias{vertical_manip}
\title{Vertically roatate 1 dim of a p-dim basis.}
\usage{
vertical_manip(manip_space, phi = 0, ...)
}
\arguments{
\item{phi}{angle changing the magnitude of `manip_var`}

\item{basis}{starting basis to rotate}

\item{manip_var}{number of the variable to manipulate}
}
\description{
Performs a vertical rotation on 1 dimension of a given p-dimensional basis.
Returns the x,y,z contribution from each dimension of the rotated basis as a (p x 3) matrix.
}
\examples{
this_manip_space <- create_manip_space(basis = basis_random(p = 5), manip_var = 2)
vertical_manip(manip_space = this_manip_space, phi = pi/3)

}
