% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{animate_gganimate}
\alias{animate_gganimate}
\title{Animate a \code{ggtour()} with the use of \code{{gganimate}}}
\usage{
animate_gganimate(
  ggtour,
  fps = 8,
  rewind = FALSE,
  start_pause = 1,
  end_pause = 1,
  knit_pdf_anim = FALSE,
  ...
)
}
\arguments{
\item{ggtour}{The return of a \code{ggtour()}and added \verb{proto_*()} functions.}

\item{fps}{Scalar number of Frames Per Second, the speed the animation should
play at.}

\item{rewind}{Whether or not the animation should play backwards,
in reverse order once reaching the end. Defaults to FALSE.}

\item{start_pause}{The duration in seconds to wait before starting the
animation. Defaults to 1 second.}

\item{end_pause}{The duration in seconds to wait after ending the animation,
before it restarts from the first frame. Defaults to 1 second.}

\item{...}{other arguments to pass to \code{gganimate::animate()}.}
}
\description{
Animates the static \code{ggtour()} and added \verb{proto_*()} functions as a
\code{{gganimate}} animation, .gif without interaction.
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
clas <- tourr::flea$species
bas <- basis_pca(dat)
mv <- manip_var_of(bas)
mt_path <- manual_tour(bas, manip_var = mv, angle = .1)

ggt <- ggtour(mt_path, dat) +
  proto_basis2d() +
  proto_data_background() +
  proto_data_points(aes_args = list(color = clas, shape = clas),
                    identity_args = list(size = 1.5, alpha = .7)) +

\dontrun{
animate_gganimate(ggtour)

(anim <- animate_gganimate(ggtour, fps = 10, rewind = TRUE,
                           start_pause = 1, end_pause = 2))

if(F) ## Example saving gganime to a .gif, may require additional setup.
  gganimate::anim_save("my_tour.gif",
                       animation = anim,
                       path = "./figures")
}
}
\seealso{
Other ggtour animator: 
\code{\link{animate_plotly}()}
}
\concept{ggtour animator}
