% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{animate_plotly}
\alias{animate_plotly}
\title{Animate a ggtour as and HTML widget via \code{{plotly}}}
\usage{
animate_plotly(ggtour, fps = 8, ...)
}
\arguments{
\item{ggtour}{A grammar of graphics tour with appended protos added.
A return from \verb{ggtour() + proto_*()}.}

\item{fps}{Scalar number of Frames Per Second, the speed the animation should
play at.}

\item{...}{Other arguments passed to
\code{\link[plotly:ggplotly]{plotly::ggplotly}}.}
}
\description{
Animates the static \code{ggtour()} and added \verb{proto_*()} functions as a
\code{{plotly}} animation, an .html widget with slider and hover tooltip showing
the row number.
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
clas <- tourr::flea$species
bas <- basis_pca(dat)
mv <- manip_var_of(bas)
mt_path <- manual_tour(bas, manip_var = mv, angle = .1)

ggt <- ggtour(mt_path, dat) +
  proto_origin() +
  proto_basis() +
  proto_point(aes_args = list(color = clas, shape = clas),
              identity_args = list(size = 1.5, alpha = .7))

\dontrun{
animate_plotly(ggtour)

(anim <- animate_plotly(ggtour, fps = 10))

if(F) ## Example saving ploty to a .html widget, may require additional setup.
  htmlwidgets::saveWidget(widget = anim, file = "./figures/my_tour.html",
                          selfcontained = TRUE)
}
}
\seealso{
\code{\link[plotly:ggplotly]{plotly::ggplotly}}

Other ggtour animator: 
\code{\link{animate_gganimate}()}
}
\concept{ggtour animator}
