% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{animate_plotly}
\alias{animate_plotly}
\title{Animate a \code{ggtour()} via \code{{plotly}}}
\usage{
animate_plotly(ggtour, fps = 8, ...)
}
\arguments{
\item{ggtour}{The return of a \code{ggtour()}and added \verb{proto_*()} functions.}

\item{fps}{Scalar number of Frames Per Second, the speed the animation should
play at.}

\item{...}{other arguments to pass to \code{gganimate::animate()}.}

\item{rewind}{Whether or not the animation should play backwards,
in reverse order once reaching the end. Defaults to FALSE.}

\item{start_pause}{The duration in seconds to wait before starting the
animation. Defaults to 1 second.}

\item{end_pause}{The duration in seconds to wait after ending the animation,
before it restarts from the first frame. Defaults to 1 second.}
}
\description{
Animates the static \code{ggtour()} and added \verb{proto_*()} functions as a \code{{plotly}}
animation, an .html widget with slider and hover tooltip showing
the row number.
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
clas <- tourr::flea$species
bas <- basis_pca(dat)
mv <- manip_var_of(bas)
mt_path <- manual_tour(bas, manip_var = mv, angle = .1)

ggt <- ggtour(mt_path, dat) +
  proto_data_background() +
  proto_basis2d() +
  proto_data_points(aes_args = list(color = clas, shape = clas),
                    identity_args = list(size = 1.5, alpha = .7))

\dontrun{
animate_plotly(ggtour)

(anim <- animate_plotly(ggtour, fps = 10))

if(F) ## Example saving ploty to a .html widget, may require additional setup.
  htmlwidgets::saveWidget(widget = anim, file = "./figures/my_tour.html",
                          selfcontained = TRUE)
}
}
\seealso{
Other ggtour animator: 
\code{\link{animate_gganimate}()}
}
\concept{ggtour animator}
