% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{basis_olda}
\alias{basis_olda}
\title{The basis of Orthogonal Linear Discriminant Analysis (OLDA)}
\usage{
basis_olda(data, class, d = 2L)

basis_olda(data, class, d = 2L)
}
\arguments{
\item{data}{Numeric matrix or data.frame of the observations, coerced to matrix.}

\item{class}{The class for each observation, coerced to a factor.}

\item{d}{Number of dimensions in the projection space.}
}
\value{
Orthogonal basis that best distinguishes the levels of \code{class}.

Orthogonal basis that best distinguishes the levels of \code{class}.
}
\description{
Returns a numeric matrix of the first \code{d} columns of \code{Rdimtools::do.olda()}.

Returns a numeric matrix of the first \code{d} columns of \code{Rdimtools::do.olda()}.
}
\examples{
basis_olda(data = wine[, 2:14], class = wine$Type)
basis_olda(data = wine[, 2:14], class = wine$Type)
}
\references{
Ye J (2005). “Characterization of a Family of Algorithms for Generalized
Discriminant Analysis on Undersampled Problems.” J. Mach. Learn. Res.,
6, 483–502. ISSN 1532-4435.

Ye J (2005). “Characterization of a Family of Algorithms for Generalized
Discriminant Analysis on Undersampled Problems.” J. Mach. Learn. Res.,
6, 483–502. ISSN 1532-4435.
}
\seealso{
\code{\link[Rdimtools]{do.olda}}

\code{\link[Rdimtools]{do.olda}}
}
