% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{basis_onpp}
\alias{basis_onpp}
\title{The basis of Orthogonal Neighborhood Preserving Projection (OLPP)}
\usage{
basis_onpp(data, d = 2, type = c("knn", sqrt(nrow(data))))
}
\arguments{
\item{data}{Numeric matrix or data.frame of the observations, coerced to matrix.}

\item{d}{Number of dimensions in the projection space.}

\item{type}{A vector specifying the neighborhood graph construction.
Expects; \code{c("knn", k)}, \code{c("enn", radius)}, or \code{c("proportion",ratio)}.
Defaults to \code{c("knn", sqrt(nrow(data)))}, nearest neighbors equal to the
square root of observations.}
}
\value{
Orthogonal matrix basis that distinguishes the levels of \code{class}
based on local and non-local variation as weighted against the neighborhood
graph.
}
\description{
Orthogonal Neighborhood Preserving Projection (ONPP) is an unsupervised
linear dimension reduction method. It constructs a weighted data graph from
LLE method. Also, it develops LPP method by preserving the structure of local
neighborhoods. For the more details on \code{type} see
\code{\link[Rdimtools:aux.graphnbd]{Rdimtools::aux.graphnbd()}}.
}
\examples{
dat_std <- scale_sd(wine[, 2:6])
basis_onpp(data = dat_std)
}
\references{
He X (2005). Locality Preserving Projections. PhD Thesis,
University of Chicago, Chicago, IL, USA.
}
\seealso{
\code{\link[Rdimtools:do.onpp]{Rdimtools::do.onpp}}

\code{\link[Rdimtools:aux.graphnbd]{Rdimtools::aux.graphnbd}} for
details on \code{type}.

Other basis identifiers: 
\code{\link{basis_guided}()},
\code{\link{basis_half_circle}()},
\code{\link{basis_odp}()},
\code{\link{basis_olda}()},
\code{\link{basis_olpp}()},
\code{\link{basis_pca}()}
}
\concept{basis identifiers}
