% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{col_of}
\alias{col_of}
\title{Return \code{col} values for a given categorical variable}
\usage{
col_of(cat, pallet_name = "Dark2")
}
\arguments{
\item{cat}{The categorical variable to return the \code{col} values for.}

\item{pallet_name}{The name of the \code{RColorBrewer} pallet to get the colors
from. Defaults to "Dark2"}
}
\value{
The integer \code{col} values for a passed categorical variable.
}
\description{
Retruns string \code{col} values for a passed categorical variable.
}
\examples{
col_of(tourr::flea$species)
}
