% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_identity_basis.r
\name{create_identity_basis}
\alias{create_identity_basis}
\title{Creates and returns an identity basis}
\usage{
create_identity_basis(p, d = 2)
}
\arguments{
\item{p}{number of dimensions of the data}

\item{d}{number of dimensions of the basis. Defaults to 2}
}
\value{
[p, d=2] identity matrix followed by rows of 0s
}
\description{
Creates a [p, d=2] identity basis; identity matrix followed by 0s
}
