% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_tour.r
\name{create_manip_space}
\alias{create_manip_space}
\title{Create a manipulation space}
\usage{
create_manip_space(basis, manip_var)
}
\arguments{
\item{basis}{A [p, d] dim orthonormal matrix.}

\item{manip_var}{Number of the dimension (numeric variable) to rotate.}
}
\value{
A [p, d+1] orthonormal matrix, the manipulation space.
}
\description{
Typically called by `manual_tour()`. Creates a [p, d] dim orthonormal matrix,
the manipulation space from the given basis right concatonated with a zero 
vector, with manip_var set to 1.
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])

rb <- tourr::basis_random(n = n col(flea_std))
create_manip_space(basis = rb, manip_var = 4)
}
