% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{create_manip_space}
\alias{create_manip_space}
\title{Create a manipulation space}
\usage{
create_manip_space(basis = basis_random(p = 5, d = 2), manip_var = 1)
}
\arguments{
\item{basis}{[p, 2] orthonormal basis}

\item{manip_var}{number of the variable to rotate}
}
\value{
[p, d+1=3] orthonormal manipulation space
}
\description{
Create an [p, d+1=3] orthonormal manipulation space from the given [p, d=2] orthonormal basis. This is used before a manipulation, creating the d+1 space to allow for the rotation.
}
\examples{
this_manip_space <- create_manip_space(basis = basis_random(p = 5), manip_var = 2)

}
