% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_data.r
\name{create_manip_space}
\alias{create_manip_space}
\title{Create a manipulation space}
\usage{
create_manip_space(basis, manip_var)
}
\arguments{
\item{basis}{A [p, d=2] dim orthonormal basis.}

\item{manip_var}{Number of the dimension (numeric variable) to rotate.}
}
\value{
manip_space, a [p, d+1=3] orthonormal manipulation space
}
\description{
Create a [p, d+1=3] dim orthonormal manipulation space from the given
basis concatonated with a zero vector, with manip_var set to 1.
}
\examples{
create_random_basis(6) -> ThisBasis
create_manip_space(ThisBasis, 4)
}
