% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{create_slides}
\alias{create_slides}
\title{Create a slideshow array of the projected bases}
\usage{
create_slides(tour, data = NULL)
}
\arguments{
\item{tour}{A [p, d, n_slides] dim array of a tour. Ie) The output of 
`manual_tour`.}

\item{data}{A [n, p] dim data to project, consisting of only numeric 
variables (for coercion into matrix).}
}
\value{
A list containing the [p, d, n_slides] dim basis slides array, and
  the [n, d, n_slides] dim data slides array.
}
\description{
Takes the result of `manual_tour()` and projects the data over the 
interpolated tour path of the reference frame.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[,1:6])

rb <- tourr::basis_random(n = ncol(flea_std))
mtour <- manual_tour(rb, manip_var = 4)
create_slides(tour = mtour, data = flea_std)
}
