% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{create_slideshow}
\alias{create_slideshow}
\title{Create a slideshow array of the projected bases}
\usage{
create_slideshow(data, m_tour, center = TRUE, scale = FALSE)
}
\arguments{
\item{data}{[n, p] dim data to project, consisting of 
only numeric variables (for coercion into matrix.)}

\item{m_tour}{the output of manual_tour(), list of projection bases by index.}

\item{center}{set the mean of the projected data to be a vector of zeros. 
This stops the data from wandering around the display window. Default=TRUE.}

\item{scale}{set the scale of the projected data to be in a standard range 
for all projections. Typically not useful, but occasionally we are only 
interested in the shape of projected data not the magnitude. Default=FALSE.}
}
\description{
Takes the result of manual_tour() and uses base graphics to view each index with delay in base grahics
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])

rb <- tourr::basis_random(n=ncol(flea_std))
mtour <- manual_tour(rb, manip_var = 4)
sshow <- create_slideshow(data = flea_std, m_tour = mtour)
}
