% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{create_slideshow}
\alias{create_slideshow}
\title{Create a slideshow array of the projected bases}
\usage{
create_slideshow(tour, data = NULL, ...)
}
\arguments{
\item{data}{A [n, p] dim data to project, consisting of 
only numeric variables (for coercion into matrix.)}

\item{m_tour}{The output of manual_tour(), a [p, d, n_slides] dim array of 
the manual tour. Containing `n_slides` interpolations varying phi.}
}
\value{
A list containing the [p, d, n_slides] dim basis slides array, and
  the [n, d, n_slides] dim data slides array.
}
\description{
Takes the result of manual_tour() and projects the data over the interpolated
tour path of the reference frame.
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])

rb <- tourr::basis_random(n = ncol(flea_std))
mtour <- manual_tour(rb, manip_var = 4)
sshow <- create_slideshow(tour = mtour, data = flea_std)
}
