% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{.bind_elements2df}
\alias{.bind_elements2df}
\title{Binds replicated elements of a list as columns of a data frame.}
\usage{
.bind_elements2df(list, df)
}
\arguments{
\item{list}{A list of arguments such as those passed in \code{aes_args} and
\code{identity_args}.}

\item{df}{A data.frame to column bind the elements of \code{list} to.}
}
\description{
Internal function. To be applied to \code{aes_args}
replicates elements to the length of the data and bind as a column.
}
\examples{
## This function is not meant for external use
}
\seealso{
Other Internal utility: 
\code{\link{.init4proto}},
\code{\link{.lapply_rep_len}()}
}
\concept{Internal utility}
