% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\docType{data}
\name{.init4proto}
\alias{.init4proto}
\title{Initialize common obj from .global \code{ggtour()} objects & test their existence}
\format{
An object of class \code{expression} of length 1.
}
\usage{
.init4proto
}
\description{
Internal expression. Creates local .objects to be commonly consumed by
spinifex proto_* functions.
}
\examples{
## This expression. is not meant for external use.
}
\seealso{
Other Internal utility: 
\code{\link{.bind_elements2df}()},
\code{\link{.lapply_rep_len}()},
\code{\link{interpolate_manual_tour}()}
}
\concept{Internal utility}
\keyword{datasets}
