% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{draw_basis}
\alias{draw_basis}
\title{draw a basis on a static ggplot}
\usage{
draw_basis(
  basis,
  map_to = data.frame(x = c(0, 1), y = c(0, 1)),
  position = c("left", "center", "right", "bottomleft", "topright", "off"),
  manip_col = "blue",
  line_size = 1,
  text_size = 5,
  label = abbreviate(rownames(basis), 3L)
)
}
\arguments{
\item{basis}{A (p*d) basis to draw. Draws the first two components.
If facet is used cbind the facet variable to a specidic facet level
(2nd example), otherwise the basis prints on all facet levels.}

\item{map_to}{A data.frame to scale the basis to.
Defaults to a unitbox; data.frame(x=c(0,1), y=c(0,1)).}

\item{position}{The position, to place the basis axes relative to the centered
data. \verb{_basis} Expects one of c("left", "center", "right", "bottomleft",
"topright", "off"), defaults to "left".}

\item{manip_col}{The color to highlight the manipulation variable with. Not
applied if the tour isn't a manual tour. Defaults to "blue".}

\item{line_size}{(2D bases only) the thickness of the lines used to make the
axes and unit circle. Defaults to 1.}

\item{text_size}{Size of the text label of the variables.}

\item{label}{The text labels of the data variables.
Defaults to the 3 character abbreviation of the rownames of the basis.}
}
\description{
Aditively draws a basis on a static ggplot.
Not a formal \code{geom}, nor does it use the setup from \code{ggtour} that
\verb{proto_*} functions expect.
}
\examples{
library(ggplot2)
dat <- scale_sd(tourr::flea[, 1:6])
bas <- basis_pca(dat)
proj <- as.data.frame(dat \%*\% bas)

ggplot() +
  geom_point(aes(PC1, PC2), proj) +
  draw_basis(bas, proj, "left")
  
## Aesthetics and facet
proj <- cbind(proj, clas = tourr::flea$species)
bas <- cbind(as.data.frame(bas), clas = levels(tourr::flea$species)[2])
ggplot() +
  facet_wrap(vars(clas)) +
  geom_point(aes(PC1, PC2, color = clas, shape = clas), proj) +
  draw_basis(bas, proj, "left")
# To repeat basis in all facet levels don't append teh facet variable.
}
