% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{ggtour}
\alias{ggtour}
\title{Prepare a new ggtour}
\usage{
ggtour(basis_array, data = NULL, angle = 0.05)
}
\arguments{
\item{basis_array}{An array of projection bases for the tour, as produced
with \code{manual_tour()} or \code{tour::save_history()}.}

\item{data}{Numeric data to project. If left NULL, will check if it data is
stored as an attribute of the the \code{basis_array}.}

\item{angle}{Target angle (in radians) for interpolation for
\code{tour::save_history()} generated \code{basis_array}. Defaults to .05.}
}
\description{
\code{ggtour()} initializes a ggplot object for a tour, to be animated with
\code{animate_plotly()} or \code{animate_ggtour()}.
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
bas <- basis_pca(dat)
mv <- manip_var_of(bas)
mt_path <- manual_tour(bas, manip_var = mv, angle = .1)

ggtour(mt_path, dat) ## Returns headless ggplot(), but required for other spinifex protos

ggt <- ggtour(mt_path, dat) +
  proto_basis2d
}
\seealso{
Other ggtour proto: 
\code{\link{proto_basis_1d}()},
\code{\link{proto_basis_2d}()},
\code{\link{proto_data_background}()},
\code{\link{proto_data_density1d_rug}()},
\code{\link{proto_data_hex}()},
\code{\link{proto_data_points}()},
\code{\link{proto_data_text}()}
}
\concept{ggtour proto}
