% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{is_orthonormal}
\alias{is_orthonormal}
\title{Test if a numeric matrix is orthonormal.}
\usage{
is_orthonormal(x, tol = 0.001)
}
\arguments{
\item{x}{Numeric matrix}

\item{tol}{Tolerance of (the sum of element-wise) floating point differences.}
}
\value{
Single logical of the orthonormal matrix of the matrix.
}
\description{
Handles more cases than tourr::is_orthonormal().
}
\examples{
is_orthonormal(tourr::basis_random(n = 6))
is_orthonormal(matrix(1:12, ncol=2))
}
