% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{is_orthornormal}
\alias{is_orthornormal}
\title{Check for orthonormality of a matrix (or basis)}
\usage{
is_orthornormal(basis)
}
\arguments{
\item{basis}{a matrix or data.frame to check for orthonormality}
}
\description{
For internal use mainly. returns TRUE or basis transposed %*% basis, would be identity matrix if basis is orthernormal
}
\examples{
basis <- matrix(c(0.707, 0, 0.707, 0, 1, 0), ncol=2, byrow=FALSE)
is_orthornormal(basis)
basis <- matrix(c(runif(6)), ncol=2, byrow=FALSE)
is_orthornormal(basis)
}
