% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{last_ggtour}
\alias{last_ggtour}
\alias{.set_last_ggtour}
\title{Retrieve/set a list from the last \code{ggtour()}, required for the use
of \verb{proto_*} functions.}
\usage{
last_ggtour()

.set_last_ggtour(ggtour_list)
}
\description{
Retrieve/set a list from the last \code{ggtour()}, required for the use
of \verb{proto_*} functions.
}
\seealso{
\code{\link[=ggtour]{ggtour()}}
}
\keyword{internal}
