% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_tour.r
\name{manual_tour}
\alias{manual_tour}
\title{Produce the series of projection bases to rotate a variable into and out 
of a projection}
\usage{
manual_tour(basis = NULL, manip_var = NULL, manip_type = "radial",
  theta = NULL, phi_min = 0, phi_max = 0.5 * pi, n_slides = 20)
}
\arguments{
\item{basis}{A [p, d] dim orthonormal matrix. Required, no default.}

\item{manip_var}{Integer column number or string exact column name of the.
variable to manipulate. Required, no default.}

\item{manip_type}{String of the type of manipulation to use. 
Defaults to "radial". Alternatively accepts "horizontal" or "vertical". 
Yields to `theta` if set. Must set either `manip_type` or `theta`.}

\item{theta}{Angle in radians of "in-plane" rotation, on the XY plane of the 
reference frame. Typically set from manip_type in `proj_data`. Supersedes 
`manip_type`. Must set either `manip_type` or `theta`.}

\item{phi_min}{Minimum value phi should move to. Phi is angle in radians of 
the "out-of-plane" rotation, the z-axis of the reference frame. 
Required, defaults to 0.}

\item{phi_max}{Maximum value phi should move to. Phi is angle in radians of 
the "out-of-plane" rotation, the z-axis of the reference frame. 
Required, defaults to 2 * pi.}

\item{n_slides}{Number of slides to create. Defaults to 20.}
}
\value{
A [p, d, n_slides] dim array of the manual tour. Containing
  `n_slides` interpolations varying phi from it's start to `phi_min`, to 
  `phi_max`, and back to start.
}
\description{
Typically called by `create_slides`. The manual tour of the `manip_var`.
Given a [p, d] orthonormal basis, creates an array of `n_slides` bases 
extending the norm of `manip_var`, via cos(phi), from `phi_max`, to 
`phi_min`, then back to the starting position (by default: from start, to 0,
to 1, to start).
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])

rb <- tourr::basis_random(n = ncol(flea_std))
manual_tour(basis = rb, manip_var = 4)
}
