% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_tour.r
\name{manual_tour}
\alias{manual_tour}
\title{Produce the series of projection bases to rotate a variable into and out
of a projection}
\usage{
manual_tour(basis = NULL, manip_var, theta = NULL, phi_min = 0,
  phi_max = 0.5 * pi, n_slides = 20)
}
\arguments{
\item{basis}{A (p, d) dim orthonormal matrix. Required, no default.}

\item{manip_var}{Integer column number or string exact column name of the.
variable to manipulate. Required, no default.}

\item{theta}{Angle in radians of "in-plane" rotation, on the XY plane of the
reference frame. Typically set from manip_type in \code{proj_data}. Supersedes
\code{manip_type}. Must set either \code{manip_type} or \code{theta}.}

\item{phi_min}{Minimum value phi should move to. Phi is angle in radians of
the "out-of-plane" rotation, the z-axis of the reference frame.
Required, defaults to 0.}

\item{phi_max}{Maximum value phi should move to. Phi is angle in radians of
the "out-of-plane" rotation, the z-axis of the reference frame.
Required, defaults to pi/2.}

\item{manip_type}{String of the type of manipulation to use.
Defaults to "radial". Alternatively accepts "horizontal" or "vertical".
Yields to \code{theta} if set. Must set either \code{manip_type} or \code{theta}.}
}
\value{
A (p, d, 4) history_array of the manual tour. The bases set for
phi_start, \code{phi_min},  \code{phi_max}, and back to phi_start. To be called by
tourr::interpolate().
}
\description{
Typically called by \code{array2af()}. An array of projections,
the manual tour of the \code{manip_var}, which is rotated from phi's starting
position to \code{phi_max}, to \code{phi_min}, and back to the start position.
}
\examples{
flea_std <- tourr::rescale(flea[,1:6])

rb <- basis_random(n = ncol(flea_std))
manual_tour(basis = rb, manip_var = 4)
}
