% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_tour.r
\name{manual_tour}
\alias{manual_tour}
\title{Produce the series of porjection bases to rotate a variable into and out of a 
projection}
\usage{
manual_tour(basis = NULL, manip_var = NULL, manip_type = "radial",
  theta = NULL, phi_min = 0, phi_max = pi, n_slides = 20)
}
\arguments{
\item{basis}{A [p, 2] dim orthonormal starting basis. Required no default.}

\item{manip_var}{Integer column number or string exact column name of the.
variable to manipulate. Required, no default.}

\item{manip_type}{String of the type of manipulation to use. 
Defaults to "radial". Alternatively accepts "horizontal" or "vertical". 
Required, supersedes theta if set.}

\item{theta}{Optional parameter, yielding to manip_type. Angle in radians 
specifying the angle between the `manip_var` and 0 on bases reference frame. 
Where the 0 is the positive side of the x-axis.}

\item{phi_min}{Angle in radians specifying the minimum extent that 
`manip_var` should extend in the z-axis w.r.t. the bases reference frame. 
Required, defaults to 0.}

\item{phi_max}{Angle in radians specifying the maximum extent that 
`manip_var` should extend in the z-axis w.r.t. the bases reference frame. 
Required, defaults to 2 * pi.}

\item{n_slides}{Number of slides to create for slideshow(). Defaults to 20.}
}
\value{
`m_tour`, a [p, d, n_slides] dim array of the manual tour transistion
in `n_slides``increments.
}
\description{
Rotates the manipulation space across `n_slides` increments to `phi_max` and
back to `phi_min`.
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])

rb <- tourr::basis_random(n=ncol(flea_std))
mtour <- manual_tour(basis = rb, manip_var = 4)
}
