% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblique_frame.r
\name{oblique_frame}
\alias{oblique_frame}
\title{Plot a single frame of a manual tour}
\usage{
oblique_frame(
  basis = NULL,
  data = NULL,
  manip_var = NULL,
  theta = 0,
  phi = 0,
  lab = NULL,
  rescale_data = FALSE,
  ...
)
}
\arguments{
\item{basis}{A (p, d) dim orthonormal numeric matrix.
If it's left null, random basis will be used.}

\item{data}{A  (n, p) dataset to project, consisting of numeric variables.}

\item{manip_var}{Number of the column/dimension to rotate.}

\item{theta}{Angle in radians of "in-plane" rotation, on the XY plane of the
reference frame. Required, no default.
If left NULL, will initialize the radial angle of the \code{manip_var}.`}

\item{phi}{Phi is angle in radians of
the "out-of-plane" rotation, the z-axis of the reference frame.
Required, no default.}

\item{lab}{Optional, labels for the reference frame of length 1 or the
number of variables used. Defaults to an abbriviation of the variables.}

\item{rescale_data}{When TRUE scales the data to between 0 and 1.
Defaults to FALSE.}

\item{...}{Optionally pass additional arguments to the \code{render_type} for
plotting options.}
}
\value{
a ggplot object of the rotated projection.
}
\description{
Projects the specified rotation as a 2D ggplot object. One static frame of
manual tour. Useful for providing uaer-guided interaction.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[,1:6])
rb <- tourr::basis_random(n = ncol(flea_std))
theta <- runif(1, 0, 2*pi)
phi <- runif(1, 0, 2*pi)

oblique_frame(data = flea_std, basis = rb, manip_var = 4, theta, phi)
}
