% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblique_frame.r
\name{oblique_frame}
\alias{oblique_frame}
\title{Plot a single frame of a manual tour}
\usage{
oblique_frame(data, basis = NULL, manip_var, theta = NULL,
  phi = NULL, manip_col = "blue", col = "black", pch = 20,
  axes = "center", rescale_data = FALSE, ...)
}
\arguments{
\item{data}{(n, p) dataset to project, consisting of numeric variables.}

\item{basis}{A (p, d) dim orthonormal numeric matrix.
If it's left null, random basis will be used.}

\item{manip_var}{Number of the column/dimension to rotate.}

\item{theta}{Angle in radians of "in-plane" rotation, on the XY plane of the
reference frame. Required, no default.
If left NULL, will initialize the radial angle of the \code{manip_var}.`}

\item{phi}{Phi is angle in radians of
the "out-of-plane" rotation, the z-axis of the reference frame.
Required, no default.}

\item{manip_col}{String of the color to highlight the \code{manip_var}.
Defaults to "blue".}

\item{col}{Color of the projected points. Defaults to "black".}

\item{pch}{Point character of the projected points. Defaults to 20.}

\item{axes}{Position of the axes: "center", "bottomleft" or "off". Defaults
to "center".}

\item{rescale_data}{When TRUE scales the data to between 0 and 1.
Defaults to FALSE.}

\item{...}{Optionally pass additional arguments to the \code{render_type} for
plotting options.}
}
\value{
a ggplot object of the rotated projection.
}
\description{
Projects the specified rotation as a 2D ggplot object. Analogous to a
slide frame of a radial tour. Used to create an oblique tour by small
changes to the rotation.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[,1:6])
rb <- tourr::basis_random(n = ncol(flea_std))
theta <- runif(1, 0, 2*pi)
phi <- runif(1, 0, 2*pi)

oblique_frame(data = flea_std, basis = rb, manip_var = 4, theta, phi)
}
