% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{pan_zoom}
\alias{pan_zoom}
\title{Pan and zoom a 2 column matrix, dataframe or scaler number}
\usage{
pan_zoom(x, x_offset = 0, y_offset = 0, scale = 1)
}
\arguments{
\item{x}{Numeric data object with 2 columns (or scaler) to scale and offset}

\item{x_offset}{Numeric value to pan in the x-direction}

\item{y_offset}{Numeric value to pan in the x-direction}

\item{scale}{Numeric value to scale/zoom the size for x}
}
\value{
Transformed numeric data object
}
\description{
Pan and zoom a 2 column matrix, dataframe or scaler number
}
\examples{
ib <- basis_init(6, 2)
pan_zoom(x = ib, x_offset = -1, y_offset = 0, scale = 2/3)
}
