% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{pch_of}
\alias{pch_of}
\title{Return \code{pch} values for a given categorical variable}
\usage{
pch_of(cat, pch_offset = 20L)
}
\arguments{
\item{cat}{The categorical variable to return the \code{pch} values for.}

\item{pch_offset}{Integer to offset the values of \code{pch} by. Defaults to 20.}
}
\value{
The integer \code{pch} values for a passed categorical variable.
}
\description{
Retruns integer \code{pch} values for a passed categorical variable.
}
\examples{
pch_of(tourr::flea$species)
}
