% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{pch_of}
\alias{pch_of}
\title{Return shape integers for a given discrete categorical variable.}
\usage{
pch_of(class)
}
\arguments{
\item{class}{The discrete categorical variable to return the shape of.}
}
\value{
Vector of integer shape values of the discrete categorical variable.
}
\description{
Return shape integers for a given discrete categorical variable.
}
\examples{
pch_of(tourr::flea$species)
}
