% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_tour.r
\name{play_tour}
\alias{play_tour}
\title{Render display of a provided tour path and data}
\usage{
play_tour(tour_path, data = NULL, ...)
}
\arguments{
\item{tour_path}{The result of tourr::save_history}

\item{data}{Optional, tour::interpolate()'s the data. p must match the 
tour_path.}

\item{...}{Optionally pass addition arguments to plotly::animation_opts().}
}
\description{
Takes the result of tourr::save_history (or manual_tour()) and optionally 
data, renders the tour and passed data into a plotly animation object.
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])
tpath <- tourr::save_history(flea_std, tourr::guided_tour(tourr::cmass))

print_tour(tour_path = tpath, data = flea_std)
}
