% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.r
\name{play_tour_path}
\alias{play_tour_path}
\title{Render display of a provided tour path}
\usage{
play_tour_path(tour_path, data = NULL, angle = 0.05,
  render_type = render_plotly, cat_var = NULL, axes = "center",
  fps = 3, ...)
}
\arguments{
\item{tour_path}{The result of \code{tourr::save_history()} or \code{manual_tour()}.}

\item{data}{Optional, number of columns must match that of \code{tour_path}.}

\item{angle}{target distance (in radians) between bases.}

\item{render_type}{Which graphics to render to. Defaults to render_plotly,
alternative use render_gganimate.}

\item{cat_var}{Categorical variable, optionally used to set the data point
color and shape.}

\item{axes}{position of the axes: center, bottomleft or off.}

\item{fps}{Frames/slides shown per second. Defaults to 3.}

\item{...}{Optionally pass additional arguments to \code{render_type}.}
}
\description{
Takes the result of \code{tourr::save_history()} or \code{manual_tour()}, interpolates
over the path and renders into a selected \code{render_type}.
}
\examples{
flea_std <- rescale(tourr::flea[,1:6])
tpath    <- save_history(flea_std, guided_tour(cmass))

play_tour_path(tour_path = tpath, data = flea_std)

play_tour_path(tour_path = tpath, data = flea_std, angle = .05, 
  render_type = render_gganimate, cat_var = flea$species, fps = 4)
}
