% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_visualize.r
\name{play_tour_path}
\alias{play_tour_path}
\title{Render display of a provided tour path}
\usage{
play_tour_path(
  tour_path = NULL,
  data = NULL,
  angle = 0.05,
  render_type = render_plotly,
  rescale_data = FALSE,
  ggtheme = ggplot2::theme_void(),
  ...
)
}
\arguments{
\item{tour_path}{The result of \code{tourr::save_history()} or \code{manual_tour()}.}

\item{data}{Optional, number of columns must match that of \code{tour_path}.}

\item{angle}{Target distance (in radians) between steps. Defaults to .05.}

\item{render_type}{Graphics to render to. Defaults to render_plotly,
alternative use render_gganimate.}

\item{rescale_data}{When TRUE scales the data to between 0 and 1.
Defaults to FALSE.}

\item{ggtheme}{Intended for passing  a ggplot2::theme().
Alternatively accepts a list of gg functions.}

\item{...}{Optionally pass additional arguments to the \code{render_type} for
projection point aesthetics; \code{geom_point(aes(...))}.}
}
\description{
Takes the result of \code{tourr::save_history()} or \code{manual_tour()}, interpolates
over the path and renders into a selected \code{render_type}.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[, 1:6])
tpath <- tourr::save_history(flea_std, tour_path = tourr::grand_tour(), max = 3)
flea_class <- tourr::flea$species

\dontrun{
play_tour_path(tour_path = tpath, data = flea_std)

play_tour_path(tour_path = tpath, data = flea_std,
               axes = "bottomleft", angle = .08,
               color = flea_class, shape = flea_class, size = 1.5,
               render_type = render_gganimate, fps = 10,
               ggtheme = list(ggplot2::theme_void(), ggplot2::ggtitle("My title")))

if (F){ ## Saving output may require additional setup
  ## Export plotly html widget
  play_tour_path(tour_path = tpath, data = flea_std,
                 axes = "left", angle = .04, fps = 6,
                 color = flea_class, shape = flea_class, size = 2,
                 render_type = render_gganimate,
                 ggtheme = list(ggplot2::theme_void(), ggplot2::ggtitle("My title")),
                 gif_path = "myOutput", gif_filename = "myRadialTour.gif")
               
  ## Export gganimate .gif
  play_tour_path(tour_path = tpath, data = flea_std,
                 axes = "left", angle = .06, fps = 10,
                 color = flea_class, shape = flea_class, size = 1.2,
                 ggtheme = list(ggplot2::theme_void(), ggplot2::ggtitle("My title")),
                 html_filename = "myRadialTour.html")
}
}
}
