% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_data_density1d_rug}
\alias{proto_data_density1d_rug}
\title{#TODO: HEADER TEXT}
\usage{
proto_data_density1d_rug(
  aes_args = list(),
  identity_args = list(),
  density_position = c("stack", "identity")
)
}
\arguments{
\item{#TODO:}{PARAMETER TEXT}
}
\description{
#TODO: paragraph TEXT
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
clas <- tourr::flea$species
bas <- basis_pca(dat)
gt_array <- save_history(dat, grand_tour(), max = 3, start = bas)

tictoc::tic("gg assign")
gg <- ggtour(gt_array, dat, map_to = "density") +
  proto_data_density1d_rug(aes_args = list(color = clas, fill = clas))
tictoc::toc()

tictoc::tic("animate")
animate_gganimate(gg) ## ~48 seconds 
tictoc::toc()
}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis_1d}()},
\code{\link{proto_basis_2d}()},
\code{\link{proto_data_background}()},
\code{\link{proto_data_hex}()},
\code{\link{proto_data_points}()},
\code{\link{proto_data_text}()}
}
\concept{ggtour proto}
