% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_data_hex}
\alias{proto_data_hex}
\title{Tour proto for data, hexagonal heatmap}
\usage{
proto_data_hex(aes_args = list(), identity_args = list(), bins = 30)
}
\arguments{
\item{aes_args}{A list of aesthetic arguments to passed to
\verb{geom_point(aes(X)}. Any mapping of the data to an aesthetic,
for example, \code{geom_point(aes(color = myCol, shape = myCol))} becomes
\code{aes_args = list(color = myCol, shape = myCol)}.}

\item{identity_args}{A list of static, identity arguments passed into
\code{geom_point()}, but outside of \code{aes()}; \code{geom_point(aes(), X)}.
Typically a single numeric for point size, alpha, or similar.}

\item{bins}{Numeric vector giving number of bins in both vertical and
horizontal directions. Defaults to 30.}
}
\description{
Adds \code{geom_hex()} of the projected data.
}
\examples{
diam_sub <- diamonds[sample(nrow(diamonds), 1000),]
dat <- scale_sd(diam_sub[, c(1, 5:6, 8:10)])
gt_path <- save_history(dat, grand_tour(), max = 3)

ggp <- ggtour(gt_path, dat) +
  proto_basis2d() +
  proto_data_hex(bins = 20)

## Doesn't work with animate_plotly()
animate_gganimate(ggp)
}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis1d}()},
\code{\link{proto_basis2d}()},
\code{\link{proto_data_density1d_rug}()},
\code{\link{proto_data_origin}()},
\code{\link{proto_data_point}()},
\code{\link{proto_data_text}()}
}
\concept{ggtour proto}
