% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_data_text}
\alias{proto_data_text}
\title{#TODO: HEADER TEXT}
\usage{
proto_data_text(
  aes_args = list(label = as.character(1:nrow(dat))),
  identity_args = list()
)
}
\arguments{
\item{#TODO:}{PARAMETER TEXT}
}
\description{
#TODO: paragraph TEXT
}
\examples{
#TODO: EXAMPLES

}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis_1d}()},
\code{\link{proto_basis_2d}()},
\code{\link{proto_data_background}()},
\code{\link{proto_data_density1d_rug}()},
\code{\link{proto_data_hex}()},
\code{\link{proto_data_points}()}
}
\concept{ggtour proto}
