% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{append_fixed_y}
\alias{append_fixed_y}
\alias{proto_density}
\alias{proto_density1d}
\title{Tour proto for data, 1D density, with rug marks}
\usage{
append_fixed_y(fixed_y)

proto_density(
  aes_args = list(),
  identity_args = list(alpha = 0.7),
  row_index = NULL,
  density_position = c("identity", "stack", "fill"),
  rug_shape = c(142, 124, NULL)
)
}
\arguments{
\item{fixed_y}{Vector of length of the data, values to fix vertical height.
Typically related to but not an explanatory variable, for instance,
predicted Y, or residuals.}

\item{aes_args}{A list of aesthetic arguments to passed to
\verb{geom_point(aes(X)}. Any mapping of the data to an aesthetic,
for example, \code{geom_point(aes(color = myCol, shape = myCol))} becomes
\code{aes_args = list(color = myCol, shape = myCol)}.}

\item{identity_args}{A list of static, identity arguments passed into
\code{geom_point()}, but outside of \code{aes()}, for instance
\code{geom_point(aes(...), size = 2, alpha = .7)} becomes
\code{identity_args = list(size = 2, alpha = .7)}.}

\item{row_index}{A numeric or logical index of rows to subset to.
Defaults to NULL, all observations.}

\item{density_position}{The \code{ggplot2} position of \code{geom_density()}. Either
c("identity", "stack"), defaults to "identity". Warning: "stack" does not
work with \code{animate_plotly()} at the moment.}

\item{rug_shape}{Numeric, the number of the shape to make rug marks.
Expects either 142, 124 or NULL, \code{|} for plotly, ggplot2 and off respectively.
defaults to 142.}
}
\description{
Adds \code{geom_density()} and \code{geom_rug()} of the projected data. Density
\code{postion = "stack"} does not work with \code{animate_plotly()}, GH issue is open.
}
\examples{
library(spinifex)
dat     <- scale_sd(penguins[, 1:4])
clas    <- penguins$species
bas     <- basis_pca(dat)
mv      <- manip_var_of(bas)
mt_path <- manual_tour(bas, manip_var = mv)

## append_fixed_y:
# Fixed y values are useful for related values that are 
# not in the X variables, _eg_ predictions or residuals of you X space.
message("don't forget to scale your fixed_y.")
dummy_y <- scale_sd(as.integer(clas) + rnorm(nrow(dat), 0, .5))
gt_path <- save_history(dat, grand_tour(d = 1), max_bases = 5)

message("append_fixed_y wants to be called early so other proto's adopt the fixed_y.")
ggt <- ggtour(gt_path, dat, angle = .3) +
  append_fixed_y(fixed_y = dummy_y) + ## insert/overwrites vertical values.
  proto_point(list(fill = clas, color = clas)) +
  proto_basis1d() +
  proto_origin()
\dontrun{
animate_plotly(ggt)
}
library(spinifex)
dat     <- scale_sd(penguins[, 1:4])
clas    <- penguins$species
gt_path <- save_history(dat, grand_tour(), max = 3)

ggt <- ggtour(gt_path, dat) +
  proto_density(aes_args = list(color = clas, fill = clas)) +
  proto_basis1d() +
  proto_origin1d()
\dontrun{
animate_plotly(ggt)
}
}
\seealso{
Other ggtour proto functions: 
\code{\link{facet_wrap_tour}()},
\code{\link{ggtour}()},
\code{\link{proto_basis}()},
\code{\link{proto_default}()},
\code{\link{proto_hex}()},
\code{\link{proto_highlight}()},
\code{\link{proto_origin}()},
\code{\link{proto_point}()},
\code{\link{proto_text}()}

Other ggtour proto functions: 
\code{\link{facet_wrap_tour}()},
\code{\link{ggtour}()},
\code{\link{proto_basis}()},
\code{\link{proto_default}()},
\code{\link{proto_hex}()},
\code{\link{proto_highlight}()},
\code{\link{proto_origin}()},
\code{\link{proto_point}()},
\code{\link{proto_text}()}
}
\concept{ggtour proto functions}
