% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_highlight}
\alias{proto_highlight}
\alias{proto_highlight_2d}
\alias{proto_highlight1d}
\title{Tour proto highlighing specified points}
\usage{
proto_highlight(
  aes_args = list(),
  identity_args = list(color = "red", size = 5, shape = 8),
  data = NULL,
  mark_initial = FALSE
)

proto_highlight1d(
  aes_args = list(),
  identity_args = list(color = "red", linetype = 2, alpha = 0.9),
  data = NULL,
  mark_initial = FALSE
)
}
\arguments{
\item{aes_args}{A list of aesthetic arguments to passed to
\verb{geom_point(aes(X)}. Any mapping of the data to an aesthetic,
for example, \code{geom_point(aes(color = myCol, shape = myCol))} becomes
\code{aes_args = list(color = myCol, shape = myCol)}.}

\item{identity_args}{A list of static, identity arguments passed into
\code{geom_point()}, but outside of \code{aes()}, for instance
\code{geom_point(aes(...), size = 2, alpha = .7)} becomes
\code{identity_args = list(size = 2, alpha = .7)}.
Typically a single numeric for point size, alpha, or similar.}

\item{data}{Numeric data to project. If left NULL, will check if it data is
stored as an attribute of the the \code{basis_array}.}

\item{mark_initial}{Logical, whether or not to leave a fainter mark at the
subset's initial position. Defaults to FALSE.}
}
\description{
A \code{geom_point} or \code{geom_segment}(*1d) call to draw attention to a subset of
points. With the addition of \code{data} argument to related \verb{proto_*} this is
mostly redundant with an addition \code{proto_point} call on a subset of the data,
but does allow for keeping the initial position with the \code{mark_initial}
argument if desired.
}
\examples{
library(spinifex)
dat     <- scale_sd(flea[, 1:6])
clas    <- flea$species
gt_path <- save_history(dat, grand_tour(), max_bases = 5)

## d = 2 case
ggt <- ggtour(gt_path, dat, angle = .3) +
  proto_default(list(color = clas, shape = clas)) +
  proto_highlight(data = dat[5,, drop=FALSE])
\dontrun{
animate_plotly(ggt)
}

## Highlight multiple observations
ggt2 <- ggtour(gt_path, dat, angle = .3) +
  proto_default(list(color = clas, shape = clas)) +
  proto_highlight(data = dat[c( 2, 6, 19), ],
                  identity_args = list(color = "blue", size = 4, shape = 4))
\dontrun{
animate_plotly(ggt2)
}
## 1D case:
gt_path1d <- save_history(dat, grand_tour(d = 1), max_bases = 3)

ggt <- ggtour(gt_path1d, dat, angle = .3) +
  proto_default1d(list(fill = clas, color = clas)) +
  proto_highlight1d(data = dat[7,, drop = FALSE ])
\dontrun{
animate_plotly(ggt)
}

## Highlight multiple observations, mark_initial defaults to off
ggt2 <- ggtour(gt_path1d, dat, angle = .3) +
  proto_default1d(list(fill = clas, color = clas)) +
  proto_highlight1d(data = dat[c(2, 6, 7), ],
                    identity_args = list(color = "green", linetype = 1))
\dontrun{
animate_plotly(ggt2)
}
}
\seealso{
Other ggtour proto functions: 
\code{\link{facet_wrap_tour}()},
\code{\link{ggtour}()},
\code{\link{proto_basis}()},
\code{\link{proto_default}()},
\code{\link{proto_density}()},
\code{\link{proto_hex}()},
\code{\link{proto_origin}()},
\code{\link{proto_point}()},
\code{\link{proto_text}()}

Other ggtour proto functions: 
\code{\link{facet_wrap_tour}()},
\code{\link{ggtour}()},
\code{\link{proto_basis}()},
\code{\link{proto_default}()},
\code{\link{proto_density}()},
\code{\link{proto_hex}()},
\code{\link{proto_origin}()},
\code{\link{proto_point}()},
\code{\link{proto_text}()}
}
\concept{ggtour proto functions}
