% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_highlight}
\alias{proto_highlight}
\alias{proto_highlight_2d}
\alias{proto_highlight1d}
\title{Tour proto highlighing specified points, will be in front or behind
\code{ggproto_point}.}
\usage{
proto_highlight(
  rownumber_index,
  aes_args = list(),
  identity_args = list(color = "red", size = 5, shape = 8),
  mark_initial = if (length(rownumber_index) == 1) TRUE else FALSE
)

proto_highlight1d(
  rownumber_index,
  aes_args = list(),
  identity_args = list(list(color = "red", size = 1.5, linetype = 2L, length = unit(1,
    "npc"), alpha = 0.5)),
  mark_initial = if (length(rownumber_index) == 1) TRUE else FALSE
)
}
\arguments{
\item{rownumber_index}{One or more integers, the row numbers of the to
highlight. Should be within 1:n, the rows of the original data.}

\item{aes_args}{A list of aesthetic arguments to passed to
\verb{geom_point(aes(X)}. Any mapping of the data to an aesthetic,
for example, \code{geom_point(aes(color = myCol, shape = myCol))} becomes
\code{aes_args = list(color = myCol, shape = myCol)}.}

\item{identity_args}{A list of static, identity arguments passed into
\code{geom_point()}, but outside of \code{aes()}, for instance
\code{geom_point(aes(...), size = 2, alpha = .7)} becomes
\code{identity_args = list(size = 2, alpha = .7)}.
Typically a single numeric for point size, alpha, or similar.}

\item{mark_initial}{Logical, whether or not to leave a fainter mark at the
subset's initial position. By default: TRUE if single row number given else
FALSE.}
}
\description{
A \code{grom_point} call to draw attention to a subset of points.
Subset the projected data frames to the specified \code{rownum_index} of the
original data.frame with specified highlighting aesthetics. Layering is
important for use with \code{proto_point}.
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
clas <- tourr::flea$species
gt_path <- tourr::save_history(dat, grand_tour(), max_bases = 5)

ggt <- ggtour(gt_path, dat) +
  proto_highlight(rownumber_index = 5) +
  proto_point()
\dontrun{
animate_plotly(ggt)
}

ggt2 <- ggtour(gt_path, dat) +
  proto_highlight(rownumber_index = c( 2, 6, 19),
                  identity_args = list(color = "blue", size = 4, shape = 2)) +
  proto_point(list(color = clas, shape = clas),
              list(size = 2, alpha = .7))
\dontrun{
animate_plotly(ggt2)
}
## 1D case:
gt_path <- tourr::save_history(dat, grand_tour(d = 1), max_bases = 3)

ggt <- ggtour(gt_path, dat) +
  proto_highlight1d(rownumber_index = 7) +
  proto_density(list(fill = clas))
\dontrun{
animate_plotly(ggt)
}

ggt2 <- ggtour(gt_path, dat) +
  proto_highlight1d(rownumber_index = c(2, 6, 7)) +
  proto_density(list(fill = clas))
\dontrun{
animate_plotly(ggt2)
}
}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis}()},
\code{\link{proto_default}()},
\code{\link{proto_density}()},
\code{\link{proto_hex}()},
\code{\link{proto_origin}()},
\code{\link{proto_point}()},
\code{\link{proto_text}()}

Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis}()},
\code{\link{proto_default}()},
\code{\link{proto_density}()},
\code{\link{proto_hex}()},
\code{\link{proto_origin}()},
\code{\link{proto_point}()},
\code{\link{proto_text}()}
}
\concept{ggtour proto}
