% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ggproto_visualize.r
\name{proto_point}
\alias{proto_point}
\alias{proto_points}
\title{Tour proto for data point}
\usage{
proto_point(aes_args = list(), identity_args = list())
}
\arguments{
\item{position}{The position, to place the basis axes relative to the centered
data. Expects one of c("left", "center", "right", "bottomleft", "topright",
"off"), defaults to "left".}

\item{manip_col}{The color to highlight the manipulation variable with. Not
applied if the tour isn't a manual tour. Defaults to "blue".}

\item{line_size}{Thickness of the lines used to make the axes and unit
circle.}

\item{text_size}{Size of the text label of the variables.}
}
\description{
Adds \code{geom_point()} of the projected data.
}
\examples{
dat <- scale_sd(tourr::flea[, 1:6])
clas <- tourr::flea$species
gt_path <- tourr::save_history(dat, grand_tour(), max_bases = 5)

ggt <- ggtour(gt_path, dat) +
  proto_point()
\dontrun{
animate_plotly(ggt)
}

ggt2 <- ggtour(mt_path, dat) +
  proto_point(list(color = clas, shape = clas),
               list(size = 2, alpha = .7))
\dontrun{
animate_plotly(ggt2)
}
}
\seealso{
Other ggtour proto: 
\code{\link{ggtour}()},
\code{\link{proto_basis}()},
\code{\link{proto_default}()},
\code{\link{proto_density}()},
\code{\link{proto_hex}()},
\code{\link{proto_origin}()},
\code{\link{proto_text}()}
}
\concept{ggtour proto}
