% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_types.r
\name{radial_manip}
\alias{radial_manip}
\title{Radially roatate 1 dim of a p-dim basis.}
\usage{
radial_manip(manip_space, phi, theta)
}
\arguments{
\item{phi}{angle changing the magnitude of `manip_var`}

\item{theta}{angle change of `manip_var`}

\item{manip_var}{number of the variable to manipulate}
}
\value{
r_space, the [3, p] x,y,z contribution of the rotated manipulation space
}
\description{
Primarily for internal use by data_proj(). Performs a radial rotation on 1 dimension of a given p-dimensional basis.
Returns the x,y,z contribution from each dimension of the rotated manipulation space [3, p].
}
