% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_render.r
\name{render_}
\alias{render_}
\title{Prepair the ggplot object before passing to either animation package.}
\usage{
render_(
  frames,
  axes = "center",
  manip_col = "blue",
  aes_args = list(),
  identity_args = list(),
  ggproto = theme_spinifex()
)
}
\arguments{
\item{frames}{The result of \code{array2df()}, a long df of the projected frames.}

\item{axes}{Position of the axes: "center", "bottomleft", "off", "left",
"right". Defaults to "center".
For more control pass a pan_zoom call with NULL x.}

\item{manip_col}{String of the color to highlight the \code{manip_var}, if used.
Defaults to "blue".}

\item{aes_args}{A list of aesthetic arguments assigned to a vector passed to
an aes() call. Anything that would be put inside of the \code{geom_point(aes(X))}.
Typically a class vector mapped to color and shape.
For example, \code{geom_point(aes(color = myClass, shape = myClass))} becomes
\code{aes_args = list(color = myClass, shape = myClass)}.}

\item{identity_args}{A list of arguments assigned to a vector passe outside
of an aes() call. Anything that would be put in \code{geom_point(aes(), X)}.
Typically a single numeric for point size, alpha, or similar
For example, \code{geom_point(aes(), size = 2, alpha = .7)} becomes
\code{identity_args = list(size = 2, alpha = .7)}.}

\item{ggproto}{A list of ggplot2 function calls.
Anything that would be "added" to ggplot(); in the case of applying a theme,
\code{ggplot() + theme_bw()} becomes \code{ggproto = list(theme_bw())}.
Intended for aesthetic ggplot2 functions (not geom_* family).}
}
\description{
Typically called by \code{render_plotly()} or \code{render_gganimate()}. Takes the
result of \code{array2df()}, and renders them into a ggplot2 object.
}
\examples{
## Setup
dat_std <- tourr::rescale(wine[, 2:14])
clas <- wine$Type
bas <- basis_lda(dat_std, clas)
mv <- manip_var_lda(dat_std, clas)
tour_array <- manual_tour(basis = bas, manip_var = mv)
tour_df <- array2df(array = tour_array, data = dat_std)

## Required arguments
render_(frames = tour_df)

## Full arguments
render_(frames = tour_df, axes = "left", manip_col = "purple"
        aes_args = list(color = clas, shape = clas),
        identity_args = list(size = .8, alpha = .7)),
        ggproto = list(theme_spinifex(),
                       ggtitle("My title"),
                       scale_color_brewer(palette = "Set2")))
}
