% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_render.r
\name{render_}
\alias{render_}
\title{Render the ggplot before the animation package}
\usage{
render_(
  slides,
  manip_col = "blue",
  col = "black",
  pch = 20,
  cex = 1,
  axes = "center",
  alpha = 1,
  graphics = "plotly",
  ...
)
}
\arguments{
\item{slides}{The result of \code{array2df()}, a long df of the projected frames.}

\item{manip_col}{String of the color to highlight the \code{manip_var}.
Defaults to "blue".}

\item{col}{Point color and fill of the data. Defaults to "black".}

\item{pch}{Point shape of the data. Defaults to 20.}

\item{cex}{Point size of the data. Defaults to 1.}

\item{axes}{Position of the axes: "center", "bottomleft", "off", "left",
"right". Defaults to "center".}

\item{alpha}{Opacity of the data points between 0 and 1. Defaults to 1.}

\item{...}{Recieves arguments from \code{play_manual_tour()} and \code{play_tour_path()}}
}
\value{
A ggplot2 object ready to be called by \code{render_plotly()} or
\code{render_gganimate()}.
}
\description{
Typically called by \code{render_plotly()} or \code{render_gganimate()}. Takes the
result of \code{array2df()}, and renders them into a ggplot2 object.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[, 1:6])

rb <- basis_random(n = ncol(flea_std))
mtour <- manual_tour(basis = rb, manip_var = 4)
sshow <- array2df(array = mtour, data = flea_std)
render_(slides = sshow)

render_(slides = sshow, axes = "bottomleft", 
        col = col_of(flea$species), pch = pch_of(flea$species), cex = 2, alpha = .5)
}
