% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_render.r
\name{render_}
\alias{render_}
\title{Prepair the ggplot object before passing to either animation package.}
\usage{
render_(
  frames,
  axes = "center",
  manip_col = "blue",
  ggproto = ggplot2::theme_void(),
  ...
)
}
\arguments{
\item{frames}{The result of \code{array2df()}, a long df of the projected frames.}

\item{axes}{Position of the axes: "center", "bottomleft", "off", "left",
"right". Defaults to "center".}

\item{manip_col}{String of the color to highlight the \code{manip_var}, if used.
Defaults to "blue".}

\item{ggproto}{Accepts a list of gg functions. Think of this as an
alternative format to \code{ggplot() + ggproto[[1]]}.
Intended for passing a \verb{ggplot2::theme_*()} and related aesthetic functions.}
}
\description{
Typically called by \code{render_plotly()} or \code{render_gganimate()}. Takes the
result of \code{array2df()}, and renders them into a ggplot2 object.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[, 1:6])
rb <- tourr::basis_random(n = ncol(flea_std))
array_frames <- manual_tour(basis = rb, manip_var = 4)
df_frames <- array2df(array = array_frames, data = flea_std)

render_(frames = df_frames)

flea_class <- tourr::flea$species
render_(frames = df_frames, axes = "bottomleft", manip_col = "purple",
        color = flea_class, shape = flea_class,
        size = 2, alpha = .5)
        
render_(frames = df_frames,
        color = flea_class, shape = flea_class,
        ggproto = list(ggplot2::theme_grey, 
                       ggplot2::scale_color_brewer("Dark2")))
}
