% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.r
\name{render_}
\alias{render_}
\title{Render the ggplot before the animation package}
\usage{
render_(slides, manip_col = "blue", cat_var = NULL, ...)
}
\arguments{
\item{slides}{The result of \code{create_slides()}.}

\item{manip_col}{String of the color to highlight the \code{manip_var}.}

\item{cat_var}{Categorical variable, optionally used to set the data point
color and shape.}

\item{...}{Optional, pass addition arguments into \code{plotly::animation_opts()}.}
}
\value{
A ggplot2 object to be called by \code{render_plotly()} or
\code{render_gganimate()}.
}
\description{
Typically called by \code{render_plotly()} or \code{render_gganimate()}. Takes the
result of \code{create_slides()}, and renders them into a ggplot2 object.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[, 1:6])

rb <- tourr::basis_random(n = ncol(flea_std))
mtour <- manual_tour(basis = rb, manip_var = 4)
sshow <- create_slides(tour = mtour, data = flea_std)
render_(slides = sshow)
}
