% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_.r
\name{render_}
\alias{render_}
\title{Render the ggplot before the animation package}
\usage{
render_(
  slides,
  manip_col = "blue",
  col = "black",
  pch = 20,
  axes = "center",
  alpha = 1,
  ...
)
}
\arguments{
\item{slides}{The result of \code{array2df()}, a long df of the projected frames.}

\item{manip_col}{String of the color to highlight the \code{manip_var}.
Defaults to "blue".}

\item{col}{Color of the projected points. Defaults to "black".}

\item{pch}{Point character of the projected points. Defaults to 20.}

\item{axes}{Position of the axes: "center", "bottomleft" or "off". Defaults
to "center".}

\item{alpha}{Opacity of the data points between 0 and 1. Defaults to 1.}

\item{...}{Recieves arguments from \code{play_manual_tour()} and \code{play_tour_path()}}
}
\value{
A ggplot2 object ready to be called by \code{render_plotly()} or
\code{render_gganimate()}.
}
\description{
Typically called by \code{render_plotly()} or \code{render_gganimate()}. Takes the
result of \code{array2df()}, and renders them into a ggplot2 object.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[, 1:6])

rb <- basis_random(n = ncol(flea_std))
mtour <- manual_tour(basis = rb, manip_var = 4)
sshow <- array2df(array = mtour, data = flea_std)
render_(slides = sshow)

render_(slides = sshow, axes = "bottomleft", 
        col = flea$species, pch = flea$species)
}
