% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{render_gg}
\alias{render_gg}
\title{Render the ggplot before the animation package}
\usage{
render_gg(slides, manip_col = "blue", cat_var = NULL, ...)
}
\arguments{
\item{slides}{The result of \code{create_slides()}.}

\item{manip_col}{String of the color to highlight the \code{manip_var}.}

\item{cat_var}{Categorical variable, optionally used to set the data point
color and shape.}

\item{...}{Optional, pass addition arguments into \code{plotly::animation_opts()}.}
}
\value{
An animation in \code{disp_type} graphics of the interpolated data and
the corrisponding reference frame.
}
\description{
Takes \code{slides}, the result of \code{create_slides()}, and renders them into a
ggplot object to be used by \code{render_plotly()} or \code{render_gganimate()};
typically not called directly.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[,1:6])

rb <- tourr::basis_random(n = ncol(flea_std))
mtour <- manual_tour(basis = rb, manip_var = 4)
sshow <- create_slides(tour = mtour, data = flea_std)
render_gg(slides = sshow)
}
