% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.r
\name{render_gganimate}
\alias{render_gganimate}
\title{Render the slides as a \emph{gganimate} animation}
\usage{
render_gganimate(slides, manip_col = "blue", cat_var = NULL,
  fps = 0.3, ...)
}
\arguments{
\item{slides}{The result of \code{create_slides()}.}

\item{manip_col}{String of the color to highlight the \code{manip_var}.}

\item{cat_var}{Categorical variable, optionally used to set the data point
color and shape.}

\item{fps}{Frames/slides shown per second. Defaults to 3.}

\item{...}{Optional, pass addition arguments into \code{plotly::animation_opts()}.}
}
\description{
Takes the result of \code{create_slides()} and renders them into a
\emph{gganimate} animation.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[,1:6])

rb <- tourr::basis_random(n = ncol(flea_std))
mtour <- manual_tour(basis = rb, manip_var = 4)
sshow <- create_slides(tour = mtour, data = flea_std)
render_gganimate(slides = sshow)
}
