% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_render.r
\name{render_gganimate}
\alias{render_gganimate}
\title{Render the frames as a \emph{gganimate} animation.}
\usage{
render_gganimate(
  fps = 30L,
  ggtheme = theme_spinifex(),
  rewind = FALSE,
  start_pause = 1L,
  end_pause = 2L,
  gif_filename = NULL,
  gif_path = NULL,
  ...
)
}
\arguments{
\item{fps}{Frames animated per second. Defaults to 30.}

\item{ggtheme}{Intended for passing  a ggplot2::theme().}

\item{rewind}{Logical, should the animation play backwards after reaching
the end? Default to FALSE.}

\item{start_pause}{Number of seconds to pause on the first frame for.
Defaults to 1.}

\item{end_pause}{Number of seconds to pause on the last frame for.
Defaults to 2.}

\item{gif_filename}{Optional, saves the animation as a GIF to this string
(without folderpath) . Defaults to NULL (no GIF saved). For more control call
\code{gganimate::anim_save()} on a return object of \code{render_gganimate()}.}

\item{gif_path}{Optional, A string of the directory path (without filename)
to save a GIF to. Defaults to NULL (current work directory).}

\item{...}{Optionally passes arguments to the projection points inside the
aesthetics; \code{geom_point(aes(...))}.}
}
\description{
Takes the result of \code{array2df()} and renders them into a
\emph{gganimate} animation.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[, 1:6])
flea_class <- tourr::flea$species
rb <- tourr::basis_random(n = ncol(flea_std))
array_frames <- manual_tour(basis = rb, manip_var = 4)
df_frames <- array2df(array = array_frames, data = flea_std)
\dontrun{
render_gganimate(frames = df_frames)

render_gganimate(frames = df_frames, axes = "bottomleft", fps = 15, rewind = TRUE,
  col = flea_class, pch = flea_class, size = 2, alpha = .6)
  
if(F){ ## Saving .gif may require additional setup
  render_gganimate(frames = df_frames, axes = "right", fps = 40, rewind = TRUE,
    col = flea_class, pch = flea_class, size = 2,
    gif_filename = "myRadialTour.gif", gif_path = "./output")
}
}
}
