% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.r
\name{render_plotly}
\alias{render_plotly}
\title{Render the slides as a \emph{plotly} animation}
\usage{
render_plotly(slides, manip_col = "blue", cat_var = NULL, fps = 3,
  ...)
}
\arguments{
\item{slides}{The result of \code{array2df()}.}

\item{manip_col}{String of the color to highlight the \code{manip_var}.}

\item{cat_var}{Categorical variable, optionally used to set the data point
color and shape.}

\item{fps}{Frames/slides shown per second. Defaults to 3.}

\item{...}{Optional, pass addition arguments into \code{plotly::animation_opts()}.}
}
\description{
Takes the result of \code{array2df()} and renders them into a
\emph{plotly} animation.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[,1:6])

rb <- tourr::basis_random(n = ncol(flea_std))
mtour <- manual_tour(basis = rb, manip_var = 4)
sshow <- array2df(tour = mtour, data = flea_std)
render_plotly(slides = sshow)

render_plotly(slides = sshow, cat_var = flea$species)
}
