% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_render.r
\name{render_plotly}
\alias{render_plotly}
\title{Render the frames as a \emph{plotly} animation.}
\usage{
render_plotly(
  fps = 8L,
  ggtheme = ggplot2::theme_void(),
  tooltip = "none",
  html_filename = NULL,
  ...
)
}
\arguments{
\item{fps}{Frames animated per second. Defaults to 30.}

\item{ggtheme}{Intended for passing  a ggplot2::theme().}

\item{tooltip}{Character vector of aesthetic mappings to show in the \code{plotly}
hover-over tooltip. Defaults to "none". "all" shows all the
aesthetic mappings. The order of variables controls the order they appear.
For example, tooltip = c("id", "frame", "x", "y", "category", "color").}

\item{html_filename}{Optional, saves the plotly object as an HTML widget to this string
(without folderpath). Defaults to NULL (not saved). For more control call
\code{htmlwidgets::saveWidget()} on a return object of \code{render_plotly()}.}

\item{...}{Optionally passes arguments to the projection points inside the
aesthetics; \code{geom_point(aes(...))}.}
}
\description{
Takes the result of \code{array2df()} and renders them into a
\emph{plotly} animation.
}
\examples{
flea_std   <- tourr::rescale(tourr::flea[, 1:6])
flea_class <- tourr::flea$species
rb <- tourr::basis_random(n = ncol(flea_std))
array_frames <- manual_tour(basis = rb, manip_var = 4)
df_frames <- array2df(array = array_frames, data = flea_std)

\dontrun{
render_plotly(frames = df_frames)

render_plotly(frames = df_frames, axes = "bottomleft", 
              col = flea_class, pch = flea_class, size = 2, alpha = .6,
              ggtheme = list(ggplot2::theme_void(), ggplot2::ggtitle("My title")),
              fps = 10, tooltip = "all")

if(F){ ## Saving .html may require additional setup
  render_plotly(frames = df_frames, axes = "right", fps = 6,
                col = flea_class, pch = flea_class, size = 1.5,
                ggtheme = list(ggplot2::theme_void(), ggplot2::ggtitle("My title")),
                html_filename = "myRadialTour.html")
}
}
}
