% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{render_slideshow}
\alias{render_slideshow}
\title{Render a slideshow of the toured data and bases}
\usage{
render_slideshow(slide_deck, disp_type = "plotly")
}
\arguments{
\item{slide_deck}{The result of create_slideshow().}

\item{disp_type}{The graphics system to use. Defaults to 'plotly'.}
}
\description{
Takes the result of create_slideshow() and renders them as a graph object of 
the `disp_type`.
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])

rb <- tourr::basis_random(n = ncol(flea_std))
mtour <- manual_tour(basis = rb, manip_var = 4)
sshow <- create_slideshow(data = flea_std, m_tour = mtour)
render_slideshow(slide_deck = sshow)
}
