% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{render_slideshow}
\alias{render_slideshow}
\title{Render a slideshow of the toured data and bases}
\usage{
render_slideshow(slide_deck, group_by = NULL, col = NULL, pch = NULL,
  disp_type = "plotly")
}
\arguments{
\item{data}{[n, p] dim data to project, consisting of 
only numeric variables (for coercion into matrix.)}
}
\description{
Takes the result of create_slideshow() and renders them as a graph object of 
the `disp_type`.
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])

rb <- tourr::basis_random(n = ncol(flea_std) )
mtour <- manual_tour(rb, manip_var = 4)
sshow <- create_slideshow(flea_std, mtour)
render_slideshow(sshow, group_by = flea$species)
}
