% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_manip_space.r
\name{rotate_manip_space}
\alias{rotate_manip_space}
\title{Rotate and return the manipulation space}
\usage{
rotate_manip_space(manip_space, theta, phi)
}
\arguments{
\item{manip_space}{[p, 3] manipulation space to be rotated}

\item{theta}{angle of rotation w.r.t. the x-y projection. Typically set from manip_type in proj_data()}

\item{phi}{angle in radians corresponding to the magnitude of manipulation}
}
\description{
This function does the manual rotation work.
Rotates [p, 3] basis matrix manipulation space by [3, 3] rotation, 
returning [p, 3] orthonormal matrix, XYZ components after rotation.
This matrix is used to make the data projection.
}
