% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rotate_manip_space}
\alias{rotate_manip_space}
\title{Rotate the manipulation space}
\usage{
rotate_manip_space(manip_space, theta = 0, phi = 0)
}
\arguments{
\item{manip_space}{[p, 3] manipulation space to be rotated}

\item{theta}{angle of rotation w.r.t. the x-y projection}

\item{phi}{angle corisponding to the magnitude of manipulation}
}
\description{
Rotates [p, 3] manipulation space by [3, 3] rotation matrix into a [p, 3] rotated space. Used by manipulations
}
