% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_manip_space.r
\name{rotate_manip_space}
\alias{rotate_manip_space}
\title{Rotate and return the manipulation space}
\usage{
rotate_manip_space(manip_space, theta, phi)
}
\arguments{
\item{manip_space}{[p, 3] manipulation space to be rotated}

\item{theta}{angle of rotation w.r.t. the x-y projection.}

\item{phi}{angle corisponding to the magnitude of manipulation}
}
\value{
r_space, a [p, 3] orthonormal rotated space
}
\description{
Primarily for internal use. Rotates [p, 3] manipulation space by [3, 3] rotation matrix into a [p, 3] rotated space. Used by manipulations
}
