% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_tour.r
\name{rotate_manip_space}
\alias{rotate_manip_space}
\title{Rotate and return the manipulation space}
\usage{
rotate_manip_space(manip_space, theta, phi)
}
\arguments{
\item{manip_space}{A [p, 3] dim manipulation space to be rotated.}

\item{theta}{Angle in radians to rotation w.r.t. the XY reference frame. 
Typically set from manip_type in proj_data().}

\item{phi}{Angle in radians corresponding to the magnitude of rotation.}
}
\value{
r_space, a [p, 3] dim rotated space.
}
\description{
This function does the manual rotation work.
Rotates a [p, 3] basis matrix manipulation space by [3, 3] rotation, 
returning [p, 3] orthonormal matrix, XYZ components after rotation.
This matrix is used to make the data projection.
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])

rb <- tourr::basis_random(n=ncol(flea_std))
msp <- create_manip_space(rb, 4) 
rotate_manip_space(msp, theta = runif(1, max = 2 * pi), 
                   phi = runif(1, max = 2 * pi) )
}
