% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_run_app.r
\name{run_app}
\alias{run_app}
\title{Runs a shiny app demonstrating manual tours}
\usage{
run_app(app_nm = "radial_tour", ...)
}
\arguments{
\item{app_nm}{name of the shiny app to run. Expects "manual_tour".}

\item{...}{Other arguments passed into \code{shiny::runApp()}.
Such as display.mode = "showcase".}
}
\value{
Runs a locally hosted shiny app.
}
\description{
Runs a local shiny app that demonstrates manual tour and comparable
traditional techniques for static projections of multivariate data sets.
}
\examples{
\dontrun{
run_app(app_nm= "radial_tour")
run_app(app_nm= "radial_tour", display.mode = "showcase")
}
}
