% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{save_history}
\alias{save_history}
\title{A wrapper muting the text byproduct of
\code{\link[tourr:save_history]{tourr::save_history}}}
\usage{
save_history(..., verbose = FALSE)
}
\arguments{
\item{...}{additional arguments passed to tour path}

\item{verbose}{Whether or not to suppress the text output byproduct from
\code{tourr::save_history()}. Defaults to FALSE.}
}
\description{
A wrapper muting the text byproduct of
\code{\link[tourr:save_history]{tourr::save_history}}
}
\examples{
tour_path <- save_history(data = wine[, 2:6], grand_tour(), max_bases = 10)
dim(tour_path)
}
\seealso{
\code{\link[tourr:save_history]{tourr::save_history}}
}
