% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{scale_colour_discrete}
\alias{scale_colour_discrete}
\alias{scale_fill_discrete}
\title{Set default color & fill for discrete variables}
\usage{
scale_colour_discrete(...)

scale_fill_discrete(...)
}
\arguments{
\item{...}{Passes arguments to ggplot2::scale_colour/fill_brewer.}
}
\description{
Masks ggplot2's default color/fill color palette for discrete variables.
}
