% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{scale_sd}
\alias{scale_sd}
\title{Centers by mean and scales by the standard deviation of each column of data.}
\usage{
scale_sd(data)
}
\arguments{
\item{data}{Numeric matrix or data.frame of the observations.}
}
\description{
Centers by mean and scales by the standard deviation of each column of data.
}
\examples{
scale_sd(data = wine[, 2:14])
}
