% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{slideshow}
\alias{slideshow}
\title{Slide show of the projected data and basis.}
\usage{
slideshow(proj_list, col = "black", pch = "")
}
\arguments{
\item{proj_list}{the output of data_proj(), list of projected data and basis by index.}

\item{col}{color of the data, 1 color or length of the data}
}
\description{
Takes the result of data_proj() and uses base graphics to view each index with delay in base grahics
}
\examples{

flea_std <- 
  apply(flea[,1:6], 2, function(x) ((x-mean(x, na.rm=TRUE))/sd(x, na.rm=TRUE)))
data <- flea_std

proj1 <- proj_data(data, manip_var=3)
slideshow(proj1)

p <- ncol(data) 
r_basis <- create_random_basis(p = p)
pch <- flea$species
col <- flea$species

proj2 <-
  proj_data(
    data = data,
    basis = r_basis,
    manip_var = 3,
    manip_type = "radial",
    phi_from = 0,
    phi_to = pi,
    n_slides = 20
  )
slideshow(proj2, col = col, pch = pch)
}
