% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{slideshow}
\alias{slideshow}
\title{Slide show of the projected data and basis.}
\usage{
slideshow(proj_list, col = "black")
}
\arguments{
\item{proj_list}{the output of data_proj(), list of projected data and basis by index.}

\item{col}{color of the data, 1 color or length of the data}
}
\description{
Takes the result of data_proj() and uses base graphics to view each index with delay in base grahics
}
