% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinifex.r
\name{spinifex}
\alias{spinifex}
\title{Render display of a manual tour of the passed data}
\usage{
spinifex(data, basis = NULL, manip_var, manip_type = "radial",
  theta = NULL, phi_min = 0, phi_max = 0.5 * pi, n_slides = 20,
  disp_type = "plotly", col = "blue", init_rescale_data = FALSE,
  init_rand_basis = FALSE, ...)
}
\arguments{
\item{data}{A [n, p] dim data to project, consisting of 
only numeric variables (for coercion into matrix).}

\item{basis}{A [p, d] dim orthonormal numeric matrix. If it's left null, a
random basis will be created.}

\item{manip_var}{Integer column number or string exact column name of the.
variable to manipulate. Required, no default.}

\item{manip_type}{String of the type of manipulation to use. 
Defaults to "radial". Alternatively accepts "horizontal" or "vertical". 
Yields to `theta` if set. Must set either `manip_type` or `theta`.}

\item{theta}{Angle in radians of "in-plane" rotation, on the XY plane of the 
reference frame. Typically set from manip_type in proj_data(). Supersedes 
`manip_type`. Must set either `manip_type` or `theta`.}

\item{phi_min}{Minimun value phi should move to. Phi is angle in radians of 
the "out-of-plane" rotation, the z-axis of the referce frame. 
Required, defaults to 0.}

\item{phi_max}{Maximum value phi should move to. Phi is angle in radians of 
the "out-of-plane" rotation, the z-axis of the referce frame. 
Required, defaults to 2 * pi.}

\item{n_slides}{The number of slide-interpolations to make for the tour.}

\item{disp_type}{The graphics system to use. Defaults to 'plotly'.}

\item{col}{String of the colo(u)r to highlight the `manip_var`.}

\item{init_rescale_data}{When TRUE will apply tourr::rescale() on the data.
Defaults to FALSE.}

\item{init_rand_basis}{When TRUE create a random basis with 
tourr::basis_random(). Yields to basis, defaults to FALSE.}

\item{...}{Optionally pass addition arguments to the `disp_type` options.}
}
\value{
An animation in `disp_type` graphics of the interpolated data and 
  the corrisponding reference frame.
}
\description{
A wrapper function for manual_tour(), create_slides(), and 
render_slideshow(). Allows the user to go from data to plotly object of 
manual tour in one function. For use with other tour paths see play_tour().
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])
rb <- tourr::basis_random(n = ncol(flea_std))

(pss <- spinifex(data = flea_std, basis = rb, manip_var = 4))

(pss2 <- spinifex(data = flea[,1:6], manip_var = 4,
                  init_rescale_data = TRUE, init_rand_basis = TRUE))
}
