% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinifex-package.r, R/wrapper.r
\docType{package}
\name{spinifex}
\alias{spinifex}
\alias{spinifex-package}
\title{spinifex}
\usage{
spinifex(data, basis = NULL, manip_var, manip_type = "radial",
  theta = NULL, phi_min = 0, phi_max = 0.5 * pi, n_slides = 20,
  manip_col = "blue", render_type = render_plotly, cat_var = NULL,
  fps = 3, init_rescale_data = FALSE)
}
\arguments{
\item{data}{(n, p) dataset to project, consisting of numeric variables.}

\item{basis}{A (p, d) dim orthonormal numeric matrix. If it's left null, a
random basis will be created.}

\item{manip_var}{Integer column number or string exact column name of the.
variable to manipulate. Required, no default.}

\item{manip_type}{String of the type of manipulation to use.
Defaults to "radial". Alternatively accepts "horizontal" or "vertical".
Yields to \code{theta} if set. Must set either \code{manip_type} or \code{theta}.}

\item{theta}{Angle in radians of "in-plane" rotation, on the XY plane of the
reference frame. Typically set from manip_type in \code{proj_data()}. Supersedes
\code{manip_type}. Must set either \code{manip_type} or \code{theta}.}

\item{phi_min}{Minimum value phi should move to. Phi is angle in radians of
the "out-of-plane" rotation, the z-axis of the reference frame. Required,
defaults to 0.}

\item{phi_max}{Maximum value phi should move to. Phi is angle in radians of
the "out-of-plane" rotation, the z-axis of the reference frame. Required,
defaults to 2 * pi.}

\item{n_slides}{The number of slide-interpolations to make for the tour.}

\item{manip_col}{String of the color to highlight the \code{manip_var}.}

\item{render_type}{Which graphics to render to. Defaults to render_plotly,
alternative use render_gganimate.}

\item{cat_var}{Categorical variable, optionally used to set the data point
color and shape.}

\item{fps}{Frames/slides shown per second. Defaults to 3.}

\item{init_rescale_data}{When TRUE will apply \code{tourr::rescale()} on the data.
Defaults to FALSE.}

\item{...}{Optionally pass addition arguments to the \code{disp_type} options.}
}
\value{
An animation in \code{disp_type} graphics of the interpolated data and
the corresponding reference frame.
}
\description{
spinifex is a package that is compatiable and extends the package \code{tourr}.
It builds the functionality for manual tours, and adds the ability for any
tour to be rendered as \code{plotly} or \code{gganimate} objects. Touring is a class
of dimensionality reducetion that orthagonally projects data from \code{p} down
to \code{d} dimensions, take a small change in \code{p}-space and interpolates again.
Manual tours manipulate a selected variable to explore the local structure,
particaularly effective after finding a interesting tour, perhaps via a
guided tour.

A wrapper function for \code{manual_tour()}, \code{create_slides()}, and
\code{render_()}. Allows the user to go from data to plotly object of
manual tour in one function. For use with other tour paths see play_tour().
}
\details{
It's main functions are:
\itemize{
\item \code{\link[=manual_tour]{manual_tour()}}
\item \code{\link[=create_slides]{create_slides()}}
\item \code{\link[=render_plotly]{render_plotly()}}
\item \code{\link[=render_gganimate]{render_gganimate()}}
\item \code{\link[=spinifex]{spinifex()}}, A wrapper function for the above.
\item \code{\link[=play_tour]{play_tour()}}, A wrapper function which tourr tour paths can be used.
}

GitHub repo: \url{https://github.com/nspyrison/spinifex}
}
\examples{
flea_std <- tourr::rescale(tourr::flea[,1:6])
rb <- tourr::basis_random(n = ncol(flea_std))

spinifex(data = flea_std, basis = rb, manip_var = 4)

spinifex(data = tourr::flea[,1:6], manip_var = 4, init_rescale_data = TRUE)
}
\seealso{
tourr (package)
}
\author{
\strong{Maintainer}: Nicholas Spyrison \email{spyrison@gmail.com}

Authors:
\itemize{
  \item Dianne Cook [thesis advisor]
}

}
