% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinifex-package.r
\docType{package}
\name{spinifex}
\alias{spinifex}
\alias{spinifex-package}
\title{spinifex}
\description{
spinifex is a package that is compatiable and extends the package \code{tourr}.
It builds the functionality for manual tours, and adds the ability for any
tour to be rendered as \code{plotly} or \code{gganimate} objects. Touring is a class
of dimensionality reducetion that orthagonally projects data from \code{p} down
to \code{d} dimensions, take a small change in \code{p}-space and interpolates again.
Manual tours manipulate a selected variable to explore the local structure,
particaularly effective after finding a interesting tour, perhaps via a
guided tour.
}
\details{
It's main functions are:
\itemize{
\item \code{\link[=manual_tour]{manual_tour()}}, Performs manual tour and returns a tour path array.
\item \code{\link[=interpolate2df]{interpolate2df()}}, Turns a tour path array into a (long) df for ggplot
graphics.
\item \code{\link[=render_plotly]{render_plotly()}}, Plots tour path df as a plotly object.
\item \code{\link[=render_gganimate]{render_gganimate()}},  Plots tour path df as a gganimate object.
\item \code{\link[=play_manual_tour]{play_manual_tour()}}, A wrapper function, applies the above for a
manual tour.
\item \code{\link[=play_tour_path]{play_tour_path()}}, A wrapper function, renders a tour path array as
\code{render_type}.
}

GitHub repo: \url{https://github.com/nspyrison/spinifex}
}
\seealso{
tourr (package)
}
\author{
\strong{Maintainer}: Nicholas Spyrison \email{spyrison@gmail.com}

Authors:
\itemize{
  \item Dianne Cook [thesis advisor]
}

}
