% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinifex-package.r, R/wrapper.r
\docType{package}
\name{spinifex}
\alias{spinifex}
\alias{spinifex-package}
\title{spinifex}
\usage{
spinifex(data, basis = NULL, manip_var, init_rescale_data = FALSE, ...)
}
\arguments{
\item{data}{A [n, p] dim data to project, consisting of 
only numeric variables (for coercion into matrix).}

\item{basis}{A [p, d] dim orthonormal numeric matrix. If it's left null, a
random basis will be created.}

\item{manip_var}{Integer column number or string exact column name of the.
variable to manipulate. Required, no default.}

\item{init_rescale_data}{When TRUE will apply `tourr::rescale`` on the data.
Defaults to FALSE.}

\item{...}{Optionally pass other arguments to `manual_tour`, `create_slides`,
`render_slideshow`, or `plotly::animation_opts`.`}
}
\value{
An animation in `disp_type` graphics of the interpolated data and 
  the corrisponding reference frame.
}
\description{
spinifex is a package that allows for manual tours. Touring is a linear 
dimensionality reducetion method that projects objects from `p` down to `d` 
dimensions. Manual tours manipulate one selected variable by a given ammount.
spinifex can also consume tour paths generated in the tourr package and
offers streamlined production of plotly graphics to animate the tour.

A wrapper function for `manual_tour`, `create_slides`, and 
`render_slideshow`. Allows the user to go from data to plotly object of 
manual tour in one function. For use with other tour paths see `play_tour`.
}
\details{
It's main functions are:
\itemize{
  \item [manual_tour()]
  \item [create_slides()]
  \item [render_slideshow()]
  \item [spinifex()], A wrapper fucntion for the above.
  \item [play_tour()], A wrapper function which tourr tour paths can be used.
}

GitHub repo: \url{https://github.com/nspyrison/spinifex}
}
\examples{
data(flea)
flea_std <- tourr::rescale(flea[,1:6])
rb <- tourr::basis_random(n = ncol(flea_std))

(pss <- spinifex(data = flea_std, basis = rb, manip_var = 4))

(pss2 <- spinifex(data = flea[,1:6], manip_var = 4,
                  init_rescale_data = TRUE, init_rand_basis = TRUE))
}
\seealso{
tourr (package)

play_tour
}
\author{
\strong{Maintainer}: Nicholas Spyrison \email{spyrison@gmail.com}

Authors:
\itemize{
  \item Dianne Cook [thesis advisor]
}

}
