% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinifex-package.r
\docType{package}
\name{spinifex}
\alias{spinifex}
\title{spinifex}
\description{
\code{spinifex} is a package that extends the package \code{tourr}.
It builds the functionality for manual tours and allows other
tours to be rendered by \code{plotly} or \code{gganimate}. Tours are a class of
dynamic linear (orthogonal) projections of numeric multivariate data from
\code{p} down to \code{d} dimensions that are viewed as an animation as \code{p}-space is
rotated. Manual tours manipulate a selected variable, exploring how they
contribute to the sensitivity of the structure in the projection. This is
particularly useful after finding an interesting basis, perhaps via a
guided tour optimizing the projection for some objective function.
}
\details{
GitHub: \url{https://github.com/nspyrison/spinifex}
}
\seealso{
\code{\link[=basis_*]{basis_*()}} \code{\link[=manual_tour]{manual_tour()}} \code{\link[=ggtour]{ggtour()}} \code{\link[=proto_default]{proto_default()}}
}
