% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_render.r
\name{theme_spinifex}
\alias{theme_spinifex}
\title{Aesthetic settings that can be applied to a ggplot object.}
\usage{
theme_spinifex()
}
\description{
A \code{ggplot2} theme (group of aesthetic settings), that can be added to a ggplot.
}
