% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{theme_spinifex}
\alias{theme_spinifex}
\title{A ggplot2 theme suggested for linear projections with spinifex.
The default value for ggproto arguments in spinifex functions.}
\usage{
theme_spinifex(...)
}
\arguments{
\item{...}{Optionally pass arguments to \code{ggplot2::theme()}.}
}
\description{
A ggplot2 theme suggested for linear projections with spinifex.
The default value for ggproto arguments in spinifex functions.
}
\examples{
theme_spinifex()

require("ggplot2")
ggplot(mtcars, aes(wt, mpg, color = as.factor(cyl))) +
  geom_point() + theme_spinifex()
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme}} for all theme options.
}
