% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{view_basis}
\alias{view_basis}
\title{View basis axes and table}
\usage{
view_basis(basis, data = NULL)
}
\arguments{
\item{basis}{[p, d=2] basis, xy contributions of the var.}

\item{data}{optional, [n, p], applies colnames to the rows of the basis.}
}
\description{
This function can be used to draw the circle with axes 
representing the projection frame. Mainly for internal use.
}
