% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_basis.r
\name{view_basis}
\alias{view_basis}
\title{View basis axes and table}
\usage{
view_basis(basis, data = NULL)
}
\arguments{
\item{basis}{[p, d=2] basis, xy contributions of the var.}

\item{data}{optional, [n, p], applies colnames to the rows of the basis.}
}
\description{
For internal use mainly
}
