% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{view_basis}
\alias{view_basis}
\title{Plot projection frame and return the axes table.}
\usage{
view_basis(basis = create_identity_basis(6), labels = paste0("V",
  1:nrow(basis)), ...)
}
\arguments{
\item{basis}{A \link{p, d} basis, XY linear combination of each dimension
(numeric variable).}

\item{labels}{Optional, character vector of \code{p} length, add name to the axes
in the reference frame, typically the variable names.}

\item{...}{Optionally pass additional arguments to \code{segments}, \code{lines}, or
\code{text}.}
}
\value{
The basis, norm, and theta, also plots the reference frame in base
graphics.
}
\description{
Uses base graphics to plot the circle with axes representing
the projection frame. Returns the corresponding table.
}
\examples{
rb <- tourr::basis_random(4, 2)
view_basis(rb)
}
