% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{view_basis}
\alias{view_basis}
\title{Plot projection frame and return the axes table.}
\usage{
view_basis(basis = create_identity_basis(6), labels = paste0("V",
  1:nrow(basis)), ...)
}
\arguments{
\item{basis}{A [p, d=2] basis, xy contributions of each dimension 
(numeric variable).}

\item{data}{Optional, of [n, p] dim, applies colnames to the rows of the 
basis.}
}
\description{
Uses base graphics to plot the circle with axes representing
the projection frame. Returns the corrisponding table.
}
\examples{
ib <- tourr::basis_random(4, 2)
view_basis(ib)
}
