% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{view_basis}
\alias{view_basis}
\title{Plot projection frame and return the axes table}
\usage{
view_basis(basis, lab = NULL, data = NULL, axes = "center")
}
\arguments{
\item{basis}{A (p, d) basis, XY linear combination of each dimension
(numeric variable).}

\item{lab}{Optional, labels for the reference frame of length 1 or the
number of variables used. By default will abbriviate data if available.}

\item{data}{Optional (n, p) data, plots xy scatterplot on the frame}

\item{axes}{Position of the axes: "center", "bottomleft" or "off". Defaults
to "center".}
}
\value{
ggplot object of the basis.
}
\description{
Uses base graphics to plot the circle with axes representing
the projection frame. Returns the corresponding table.
}
\examples{
rb <- basis_random(4, 2)
view_basis(basis = rb)

flea_std <- tourr::rescale(tourr::flea[, 1:4])
view_basis(basis = rb, data = flea_std, axes = "bottomleft")
}
