% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_visualize.r
\name{view_frame}
\alias{view_frame}
\title{Plot a single frame of a manual tour}
\usage{
view_frame(
  basis = NULL,
  data = NULL,
  manip_var = NULL,
  theta = 0L,
  phi = 0L,
  lab = NULL,
  rescale_data = FALSE,
  ggproto = theme_spinifex(),
  ...
)
}
\arguments{
\item{basis}{A (p, d) dim orthonormal numeric matrix.
Defaults to NULL, giving a random basis.}

\item{data}{A (n, p) dataset to project, consisting of numeric variables.}

\item{manip_var}{Optional, number of the variable to rotate.
If NULL, theta and phi must be 0 as is no manip space to rotate.}

\item{theta}{Angle in radians of "in-projection plane" rotation,
on the xy plane of the reference frame. Defaults to 0, no rotation.}

\item{phi}{Angle in radians of the "out-of-projection plane" rotation, into
the z-direction of the axes. Defaults to 0, no rotation.}

\item{lab}{Optionally, provide a character vector of length p (or 1)
to label the variable contributions to the axes, Default NULL,
results in a 3 character abbreviation of the variable names.}

\item{rescale_data}{When TRUE scales the data to between 0 and 1.
Defaults to FALSE.}

\item{ggproto}{Accepts a list of gg functions. Think of this as an
alternative format to \code{ggplot() + ggproto[[1]]}.
Intended for passing a \verb{ggplot2::theme_*()} and related aesthetic functions.}

\item{...}{Optionally pass additional arguments to the \code{render_type} for
projection point aesthetics; \code{geom_point(aes(...))}.}
}
\value{
A ggplot object of the rotated projection.
}
\description{
Projects the specified rotation as a 2D ggplot object. One static frame of
manual tour. Useful for providing user-guided interaction.
}
\examples{
dat <- tourr::flea[, 1:6]
bas_pca <- stats::prcomp(dat)$rotation[, 1L:2L]
mvar <- which(abs(bas_pca[, 1]) == max(abs(bas_pca[, 1]))) ## Larget var in PC1
view_frame(bas_pca, manip_var = mvar)

rb     <- tourr::basis_random(n = ncol(dat))
rtheta <- runif(1, 0, 2 * pi)
rphi   <- runif(1, 0, 2 * pi)
view_frame(basis = rb, data = dat, manip_var = 4,
           theta = rtheta, phi = rphi, lab = paste0("MyNm", 3:8), 
           rescale_data = TRUE,
           ggproto = list(ggplot2::theme_void(), ggplot2::ggtitle("My title")))
}
