% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_visualize.r
\name{view_manip_space}
\alias{view_manip_space}
\title{Plot projection frame and return the axes table.}
\usage{
view_manip_space(
  basis,
  manip_var,
  tilt = 0.1 * pi,
  lab = paste0("V", 1:nrow(basis)),
  manip_col = "blue",
  manip_sp_col = "red",
  line_size = 1,
  text_size = 5,
  ggproto = list(theme_spinifex())
)
}
\arguments{
\item{basis}{A (p, d) orthonormal numeric matrix.
The linear combination the original variables contribute to projection space.
Required, no default.}

\item{manip_var}{Number of the column/dimension to rotate.}

\item{tilt}{angle in radians to rotate the projection plane.
Defaults to .1 * pi.}

\item{lab}{Optional, character vector of \code{p} length, add name to the axes
in the reference frame, typically the variable names.}

\item{manip_col}{String of the color to highlight the \code{manip_var}.}

\item{manip_sp_col}{Color to illustrate the z direction, orthogonal to the
projection plane.}

\item{line_size}{The size of the lines of the unit circle and variable
contributions of the basis. Defaults to 1.}

\item{text_size}{The size of the text labels of the variable
contributions of the basis. Defaults to 5.}

\item{ggproto}{Accepts a list of gg functions. Think of this as an
alternative format to \code{ggplot() + ggproto[[1]]}.
Intended for passing a \verb{ggplot2::theme_*()} and related aesthetic functions.}
}
\value{
ggplot object of the basis.
}
\description{
Uses base graphics to plot the circle with axes representing
the projection frame. Returns the corresponding table.
}
\examples{
flea_std <- tourr::rescale(tourr::flea[, 1:6])
rb <- tourr::basis_random(ncol(flea_std))

view_manip_space(basis = rb, manip_var = 4)

view_manip_space(basis = rb, manip_var = 1,
                 tilt = 2/12 * pi, lab = paste0("MyLabs", 1:nrow(basis))
                 manip_col = "purple", manip_sp_col = "orange", 
                 ggproto = list(ggplot2::theme_void(), ggplot2::ggtitle("My title")))
}
